% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Access_reanalysis_cerra_single_levels.r
\name{Access_reanalysis_cerra_single_levels}
\alias{Access_reanalysis_cerra_single_levels}
\title{Access CERRA Reanalysis Data for Europe}
\usage{
Access_reanalysis_cerra_single_levels(
  variable,
  levelType,
  soilLayer,
  dataType,
  productType,
  leadTimeHour,
  dateStart,
  dateStop,
  timeZone = "UTC",
  fileName,
  compression = NA,
  writeFile = TRUE,
  removeTemporary = TRUE,
  apiUser,
  apiKey,
  tChunkSize = 6000,
  closeConnections = TRUE
)
}
\arguments{
\item{variable}{Character. An overview of CERRA Reanalysis Data on Single Levels variables can be obtained with \code{Discovery_Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{levelType}{Character. One of "Surface or atmosphere" or "Soil". Setting aligns with choice of variable. Options for \code{levelType} per variable can be found in the output returned by \code{Discovery_Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{soilLayer}{Optional, Character. One of either "" (empty string) or "Top layer". Setting aligns with choice of variable. Options for \code{soilLayer} per variable can be found in the output returned by \code{Discovery_Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{dataType}{Character. One of either "Ensemble members" or "Reanalysis". Setting aligns with choice of variable. Options for \code{dataType} per variable can be found in the output returned by \code{Discovery_Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{productType}{Character. One of either "Analysis" or "Forecast". Setting aligns with choice of variable. Options for \code{productType} per variable can be found in the output returned by \code{Discovery_Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{leadTimeHour}{Optional, Integer. Lead time of reanalysis. CERRA Reanalysis Data on Single Levels leadtimes can be obtained with \code{Discovery_QuickFacts("CDS_reanalysis-cerra-single-levels")$leadtime}. This argument only takes effect when the \code{productType} argument is "Forecast".}

\item{dateStart}{Character. "YYYY-MM-DD HH" date at which to start time series of downloaded data. Data is available daily at hours 00, 03, 06, 09, 12, 15, 18, and 21 (in UTC time zone).}

\item{dateStop}{Character. "YYYY-MM-DD HH" date at which to stop time series of downloaded data. Data is available daily at hours 00, 03, 06, 09, 12, 15, 18, and 21 (in UTC time zone).}

\item{timeZone}{Optional, Character. Time zone in which to represent and evaluate time dimension of data. See the output of OlsonNames() for a full overview of supported specifications. Default is UTC.}

\item{fileName}{Character. A file name for the produced file, including path.}

\item{compression}{Optional, Integer. Compression level between 1 to 9 applied to final .nc file. Same as compression argument in terra::writeCDF(). Defaults to NA.}

\item{writeFile}{Optional, Logical. Whether to write final SpatRaster to disk as an .nc or to return information from memory. Setting to FALSE will prohibit removal of temporary files. Defaults to TRUE.}

\item{removeTemporary}{Optional, Logical. Whether to delete temporary files after completion. Defaults to TRUE.}

\item{apiUser}{Character. ECMWF cds user number.}

\item{apiKey}{Character. ECMWF cds API key.}

\item{tChunkSize}{Optional, Numeric. Number of layers to bundle in each individual download. Default is 6000 to adhere to most restrictive CDS limits: https://cds.climate.copernicus.eu/live/limits.}

\item{closeConnections}{Optional, Logical. Whether to close all connections at the end of function execution. When executing this function often after another, this can be useful to avoid errors. Defaults to TRUE.}
}
\value{
A SpatRaster object containing the downloaded data, and a file in the specified directory. The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string which to use for in-line citation of the data product obtained.
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call.
}

\strong{ATTENTION:} If data is loaded again from disk at a later point with a different function, take note that the time zone will have to be set anew and existing time parameters in the .nc contents will need to be converted to the desired time zone. Likewise, citation and download-call metadata will not be loaded properly from a .nc when loading data through a different function. \code{ClimHub} download funtions handle these .nc specific issues when loading .nc files created previously with \code{ClimHub} download funtions from disk automatically.
}
\description{
Downloads and processes CERRA sub-daily regional reanalysis data for Europe on single levels from 1984 to present via the \href{https://cds.climate.copernicus.eu/#!/home}{Climate Data Store (CDS)} hosted by the \href{https://cds.climate.copernicus.eu/about-c3s}{Copernicus Climate Change Service (C3S)}. By default, this function breaks down download calls into intervals to avoid failing queries, downloads data from \href{https://www.ecmwf.int/}{ECMWF} servers according to user-specification, and fuses the downloaded files together. The actual download time depends on ECMWF download queues. Users need an \href{https://cds.climate.copernicus.eu/api-how-to}{API key} and must accept terms and conditions for the specific queried dataset(s).
}
\examples{
\dontrun{
CERRA_analysis_rast <- Access_reanalysis_cerra_single_levels(
    # CDS API call specification
    variable = "2m temperature",
    levelType = "Surface or atmosphere",
    soilLayer = "",
    dataType = "Reanalysis",
    productType = "Analysis",
    leadTimeHour = "",
    # time-window, default set to range of dataset-type
    dateStart = "1993-01-01 01:00",
    dateStop = "1995-12-31 22:00",
    timeZone = "CET",
    # file naming and output
    fileName = file.path(getwd(), "CERRA_Analysis.nc"),
    compression = NA,
    writeFile = FALSE,
    removeTemporary = FALSE,
    # API credentials
    apiUser = apiUser,
    apiKey = apiKey,
    # Quality of life and efficiency settings
    tChunkSize = 6000,
    closeConnections = TRUE
)

CERRA_ensembles_rast <- Access_reanalysis_cerra_single_levels(
    # CDS API call specification
    variable = "2m relative humidity",
    levelType = "Surface or atmosphere",
    soilLayer = "",
    dataType = "Ensemble members",
    productType = "Analysis",
    leadTimeHour = "",
    # time-window, default set to range of dataset-type
    dateStart = "1993-01-01 01:00",
    dateStop = "1993-01-31 22:00",
    timeZone = "CET",
    # file naming and output
    fileName = "CERRA_Ensembles.nc",
    compression = NA,
    writeFile = TRUE,
    removeTemporary = TRUE,
    # API credentials
    apiUser = apiUser,
    apiKey = apiKey,
    # Quality of life and efficiency settings
    tChunkSize = 6000,
    closeConnections = TRUE
)
}
}
\author{
Erik Kusch
}
