% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anomaly.R
\name{anomaly}
\alias{anomaly}
\title{Compute anomalies from climate projection data}
\usage{
anomaly(var, baseline, era, period = "month", absolute = TRUE, progress = NULL)
}
\arguments{
\item{var}{A \code{CFVariable} instance to calculate the anomalies for. The \code{var}
must have a temporal axis.}

\item{baseline}{The baseline against which to calculate the anomalies. A
range of years (\code{integer}) present in \code{var}.}

\item{era}{The range of years over which to calculate the anomaly, or a list
thereof. If a list, any names will be used in the output.}

\item{period}{Optionally, the period over which each anomaly is to be
calculated, either of "day", "dekad" (10-day period), "month" (default),
"season", "quarter" or "year". Data in \code{var} is averaged to this period
before the anomaly is calculated so the data in \code{var} has to have the same
or higher resolution than this \code{period} specifies.}

\item{absolute}{Optional logical to indicate if the anomaly is to be an
absolute value in units of argument \code{var} or relative to the baseline.
Default is \code{TRUE}.}

\item{progress}{Optional progress bar. One call to this function makes 4
ticks.}
}
\value{
A \code{CFVariable} with the anomaly data, or a (named) list according to
the \code{era} argument. If either or both of the \code{baseline} or \code{era} range of
years are outside of the temporal range of \code{var}, \code{NULL} is returned.
}
\description{
Compute the anomalies of a data variable for a chosen era (range
of years) relative to a baseline, as an absolute anomaly or relative to the
baseline. The anomaly is calculated over all axes other than the temporal
axis.
}
\details{
The anomaly is calculated by differencing values from the \code{era} by
those of the \code{baseline}. Values are first averaged over the \code{period} and
then the anomaly is computed for each \code{period} separately from the "era"
and "baseline" values. In the output the temporal axis will be replaced
with a climatological axis for each of the periods.

With relative anomalies there is an asymptotic behaviour close to 0 values
for the baseline. The user is advised to determine that the resulting
anomalies make sense in the context of the data.
}
\examples{
\dontrun{
# pr is a CFVariable instance with a long time series of precipitation values
anom <- anomaly(pr, 1991:2020, 2041:2060)
}
}
\author{
Patrick Van Laake
}
