% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_Citation}
\alias{Discovery_Citation}
\title{Get citation for dataset}
\usage{
Discovery_Citation(dataSet = "NULL")
}
\arguments{
\item{dataSet}{Character. Name of dataset. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Discovery_Library()}}.}
}
\value{
Character. Citation string for dataset.
}
\description{
Retrieve and return the proper citation for a specific dataset.
}
\examples{
Discovery_Citation(dataSet = "NORA3")

}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Read}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_QuickFacts}}, \code{\link{Discovery_DOI}}.
}
\author{
Erik Kusch
}
