% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_DataSet.Check}
\alias{Discovery_DataSet.Check}
\title{Check dataSet string validity}
\usage{
Discovery_DataSet.Check(dataSet)
}
\arguments{
\item{dataSet}{Character. Name of data set. Usually a set of words separated by dashes. See possible dataSets by calling \code{\link{Discovery_Library()}}.}
}
\value{
Throws an error message if user-provided dataSet string is not recognised in metadata library.
}
\description{
Check if dataSet string is resolvable in metadata library.
}
\examples{
Discovery_DataSet.Check(dataSet = "NORA3")
# Discovery_DataSet.Check(dataSet = "NULL")
}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_DOI}}, \code{\link{Discovery_QuickFacts}}.
}
\author{
Erik Kusch
}
