% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_DOI}
\alias{Discovery_DOI}
\title{Get DOI of dataset}
\usage{
Discovery_DOI(dataSet = "NULL")
}
\arguments{
\item{dataSet}{Character. Name of dataset. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Discovery_Library()}}.}
}
\value{
Character. DOI string for dataset.
}
\description{
Retrieve and return the DOI (Digital Object Identifier) for a specific dataset for easy citation.
}
\examples{
Discovery_DOI(dataSet = "NORA3")

}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Read}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_QuickFacts}}, \code{\link{Discovery_Citation}}.
}
\author{
Erik Kusch
}
