% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_Library}
\alias{Discovery_Library}
\title{List supported datasets}
\usage{
Discovery_Library()
}
\value{
Character vector. Names of supported datasets.
}
\description{
Provides an overview of all datasets for which metadata files are present in the package.
}
\examples{
Discovery_Library()

}
\seealso{
\code{\link{Discovery_Read}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_DOI}}, \code{\link{Discovery_QuickFacts}}, \code{\link{Discovery_Citation}}.
}
\author{
Erik Kusch
}
