% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_QuickFacts}
\alias{Discovery_QuickFacts}
\title{Quick facts about dataset}
\usage{
Discovery_QuickFacts(dataSet = "NULL")
}
\arguments{
\item{dataSet}{Character. Name of dataset. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Discovery_Library()}}.}
}
\value{
List. Contains information on queried dataset in standardised way:
\itemize{
\item{DataSet}{data set string}.
\item{Type}{character, supported types of the data set}.
\item{URL}{character, url of CDS webpage corresponding to data set}.
\item{Description}{character, plain text description of data set scraped from CDS}.
\item{TResolution}{character, base temporal resolution of each layer in data set}.
\item{TStep}{numeric, vector of time step between layers in data set corresponding to Type}.
\item{TStart}{POSIXct, date and time at which first layer is available}.
\item{TEnd}{POSIXct or character, date and time at which first layer is available}.
\item{Projection}{crs of data set}.
\item{SpatialResolution}{numeric, resolution of data set in space in degrees}.
\item{CDSArguments}{list, required arguments for CDS call beyond standard arguments and also reporting default/options for common CDS query arguments}.
}
}
\description{
Retrieves and returns a short overview of dataset characteristics, including supported types, extent, time frames, and required arguments.
}
\examples{
Discovery_QuickFacts("NORA3")

}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Read}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_DOI}}, \code{\link{Discovery_Citation}}.
}
\author{
Erik Kusch
}
