% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_Read}
\alias{Discovery_Read}
\title{Data set overview}
\usage{
Discovery_Read(dataSet = "NULL")
}
\arguments{
\item{dataSet}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Discovery_Library()}}.}
}
\value{
List. Contains information of data set, type, variables, resolution, citation, etc.
}
\description{
Read and return metadata for specific data set.
}
\examples{
Discovery_Read(dataSet = "NORA3")

}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Variables}}, \code{\link{Discovery_DOI}}, \code{\link{Discovery_QuickFacts}}, \code{\link{Discovery_Citation}}.
}
\author{
Erik Kusch
}
