% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovery_Functionality.r
\name{Discovery_Variables}
\alias{Discovery_Variables}
\title{List dataset variables}
\usage{
Discovery_Variables(dataSet = "NULL")
}
\arguments{
\item{dataSet}{Character. Name of dataset. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Discovery_Library()}}.}
}
\value{
Data frame. Contains at least 3 columns:
\itemize{
\item{name}{Variable clear name and what is used to query a download}
\item{unit}{Unit of measurement}
\item{issues}{Potential known issues with this variable}
}
May contain additional columns mostly used for backend operations.
}
\description{
Reads and returns an overview of variables available for a specific dataset.
}
\examples{
Discovery_Variables(dataSet = "NORA3")

}
\seealso{
\code{\link{Discovery_Library}}, \code{\link{Discovery_Read}}, \code{\link{Discovery_DOI}}, \code{\link{Discovery_QuickFacts}}, \code{\link{Discovery_Citation}}.
}
\author{
Erik Kusch
}
