% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.KlimaiNorge2100.r
\name{Download.KlimaiNorge2100}
\alias{Download.KlimaiNorge2100}
\title{Downloading Klima i Norge 2100 Data from the Norwegian Meteorological Instite}
\usage{
Download.KlimaiNorge2100(
  Variable,
  DateStart,
  DateStop,
  Model,
  Scenario,
  Cores = 1,
  Dir = getwd(),
  FileName,
  Compression = NA,
  RemoveTemporary = TRUE,
  WriteFile = TRUE
)
}
\arguments{
\item{Variable}{Character. An overview of Klima i Norge variables can be obtained with \code{Meta.Variables(dataset = "KlimaiNorge2100")}.}

\item{DateStart}{Character. "YYYY-MM-DD" date at which to start time series of downloaded data. Data is available daily at hourly intervals.}

\item{DateStop}{Character. "YYYY-MM-DD" date at which to stop time series of downloaded data. Data is available daily at hourly intervals.}

\item{Model}{Character. An overview of climate models from which data can be obtained can be obtained with \code{Meta.QuickFacts("KlimaiNorge2100")$models}.}

\item{Scenario}{Character. An overview of climate models from which data can be obtained can be obtained with \code{Meta.QuickFacts("KlimaiNorge2100")$scenarios}. Note that this choice only affects data post-dating 2005-12-31.}

\item{Cores}{Integer. How many cores to use for parallel downloads. Default NULL defines no parallelisation.}

\item{Dir}{Character/Directory Pointer. Directory specifying where to download data to.}

\item{FileName}{Character. A file name for the produced file.}

\item{Compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}

\item{RemoveTemporary}{Logical. Whether to delete temporary files after completion.}

\item{WriteFile}{Logical. Whether to write final SpatRaster to disk as an .nc or to return information from memory. Note that setting WriteFile = FALSE will prohibit removal of temporary files.}
}
\value{
A SpatRaster object containing the downloaded data, and a file in the specified directory. The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string which to use for in-line citation of the data product obtained.
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call
}
}
\description{
This function is used to obtain the \href{https://www.miljodirektoratet.no/publikasjoner/2015/september-2015/klima-i-norge-2100/}{Klima_i_Norge_2100} data product hosted through \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/catalog.html}{thredds.met.no}. Specifically, this function makes available the following datasets:
\enumerate{
\item \href{https://publikasjoner.nve.no/rapport/2016/rapport2016_59.pdf}{Gridded 1 x 1 km climate and hydrological projections for Norway} data contained within \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/utgave2015/catalog.html}{utgave 2015}.
}
}
\examples{
\dontrun{
Download.KlimaiNorge2100(
    Variable = "Maximum Air Temperature",
    DateStart = "2004-08-01",
    DateStop = "2006-09-17",
    Model = "CNRM_CCLM",
    Scenario = "rcp85",
    Cores = 1,
    Dir = getwd(),
    FileName = "KlimaiNorge2100",
    Compression = 9,
    RemoveTemporary = TRUE
)
}
}
