% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.reanalysis_cerra_single_levels.r
\name{Download.reanalysis_cerra_single_levels}
\alias{Download.reanalysis_cerra_single_levels}
\title{Downloading CERRA sub-daily regional reanalysis data for Europe on single levels from 1984 to present Data from ECMWF Climate Data Store}
\usage{
Download.reanalysis_cerra_single_levels(
  Variable,
  LevelType,
  SoilLayer,
  DataType,
  ProductType,
  LeadTimeHour,
  DateStart,
  DateStop,
  TimeZone = "UTC",
  Dir = getwd(),
  FileName,
  Compression = NA,
  WriteFile = TRUE,
  RemoveTemporary = TRUE,
  API_User,
  API_Key,
  TChunkSize = 6000,
  closeConnections = TRUE
)
}
\arguments{
\item{Variable}{Character. An overview of CERRA Reanalysis Data on Single Levels variables can be obtained with \code{Meta.Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{LevelType}{Character. One of "Surface or atmosphere" or "Soil". Setting aligns with choice of variable. Options for \code{LevelType} per variable can be found in the output returned by \code{Meta.Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{SoilLayer}{Character. One of either "" (empty string) or "Top layer". Setting aligns with choice of variable. Options for \code{SoilLayer} per variable can be found in the output returned by \code{Meta.Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{DataType}{Character. One of either "Ensemble members" or "Reanalysis". Setting aligns with choice of variable. Options for \code{DataType} per variable can be found in the output returned by \code{Meta.Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{ProductType}{Character. One of either "Analysis" or "Forecast". Setting aligns with choice of variable. Options for \code{ProductType} per variable can be found in the output returned by \code{Meta.Variables("CDS_reanalysis-cerra-single-levels")}.}

\item{DateStart}{Character. "YYYY-MM-DD HH" date at which to start time series of downloaded data. Data is available daily at hours 00, 03, 06, 09, 12, 15, 18, and 21 (in UTC time zone).}

\item{DateStop}{Character. "YYYY-MM-DD HH" date at which to stop time series of downloaded data. Data is available daily at hours 00, 03, 06, 09, 12, 15, 18, and 21 (in UTC time zone).}

\item{TimeZone}{Character. Time zone in which to represent and evaluate time dimension of data. See the output of OlsonNames() for a full overview of supported specifications. Default is UTC.}

\item{Dir}{Character/Directory Pointer. Directory specifying where to download data to.}

\item{FileName}{Character. A file name for the produced file.}

\item{Compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}

\item{WriteFile}{Logical. Whether to write final SpatRaster to disk as an .nc or to return information from memory. Note that setting WriteFile = FALSE will prohibit removal of}

\item{RemoveTemporary}{Logical. Whether to delete temporary files after completion.}

\item{API_User}{Character; ECMWF cds user number.}

\item{API_Key}{Character; ECMWF cds API key.}

\item{TChunkSize}{Numeric. Number of layers to bundle in each individual download. Default is 6000 to adhere to most restrictive CDS limits: https://cds.climate.copernicus.eu/live/limits.}

\item{closeConnections}{Logical. Whether to close all connections at the end of function execution. When executing this function often after another, this can be very useful to avoid errors.}

\item{LeadtimeHour}{Integer. Lead time of reanalysis. CERRA Reanalysis Data on Single Levels leadtimes can be obtained with \code{Meta.QuickFacts("CDS_reanalysis-cerra-single-levels")$leadtime}. This argument only takes effect when the \code{ProductType} argument is "Forecast".}
}
\value{
A SpatRaster object containing the downloaded data, and a file in the specified directory. The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string which to use for in-line citation of the data product obtained.
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call.
}

\strong{ATTENTION:} If data is loaded again from disk at a later point with a different function, take note that the time zone will have to be set anew and existing time parameters in the .nc contents will need to be converted to the desired time zone. Likewise, citation and download-call metadata will not be loaded properly from a .nc when loading data through a different function. \code{ClimHub} download funtions handle these .nc specific issues when loading .nc files created previously with \code{ClimHub} download funtions from disk automatically.
}
\description{
This function is used to obtain data from the \href{https://cds.climate.copernicus.eu/datasets/reanalysis-cerra-single-levels?tab=overview}{CERRA sub-daily regional reanalysis data for Europe on single levels from 1984 to present} via the \href{https://cds.climate.copernicus.eu/#!/home}{Climate Data Store (CDS)} hosted by the \href{https://cds.climate.copernicus.eu/about-c3s}{Copernicus Climate Change Service (C3S)}. By default, this function breaks down download calls into intervals so as to avoid submitting queries which fail, downloads queried data from \href{https://www.ecmwf.int/}{ECMWF} servers according to user-specification, and fuses the downloaded files together according to user-demands. The actual time to download is dependent on ECMWF download queues. Users need an \href{https://cds.climate.copernicus.eu/api-how-to}{API key} for download staging and accept terms and conditions for the specific queried dataset(s).
}
\examples{
\dontrun{
CERRA_analysis_rast <- Download.reanalysis_cerra_single_levels(
    # CDS API call specification
    Variable = "2m temperature",
    LevelType = "Surface or atmosphere",
    SoilLayer = "",
    DataType = "Reanalysis",
    ProductType = "Analysis",
    LeadTimeHour = "",
    # time-window, default set to range of dataset-type
    DateStart = "1993-01-01 01:00",
    DateStop = "1995-12-31 22:00",
    TimeZone = "CET",
    # file naming and output
    Dir = getwd(),
    FileName = "CERRA_Analsysis.nc",
    Compression = NA,
    WriteFile = FALSE,
    RemoveTemporary = FALSE,
    # API credentials
    API_User = API_User,
    API_Key = API_Key,
    # Quality of life and efficiency settings
    TChunkSize = 6000,
    closeConnections = TRUE
)

CERRA_ensembles_rast <- Download.reanalysis_cerra_single_levels(
    # CDS API call specification
    Variable = "2m relative humidity",
    LevelType = "Surface or atmosphere",
    SoilLayer = "",
    DataType = "Ensemble members",
    ProductType = "Analysis",
    LeadTimeHour = "",
    # time-window, default set to range of dataset-type
    DateStart = "1993-01-01 01:00",
    DateStop = "1993-01-31 22:00",
    TimeZone = "CET",
    # file naming and output
    Dir = getwd(),
    FileName = "CERRA_Ensembles.nc",
    Compression = NA,
    WriteFile = TRUE,
    RemoveTemporary = TRUE,
    # API credentials
    API_User = API_User,
    API_Key = API_Key,
    # Quality of life and efficiency settings
    TChunkSize = 6000,
    closeConnections = TRUE
)
}
}
