% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handle.ListLayers.r
\name{Handle.ListLayers}
\alias{Handle.ListLayers}
\title{Layer-wise application on SpatRasters in a list}
\usage{
Handle.ListLayers(Rast_ls, FUN = mean)
}
\arguments{
\item{Rast_ls}{List. List of SpatRasters whose layers to compute with each other. All SpatRasters need to have the same layer count.}

\item{FUN}{Function. Function by which to summarise/compute SpatRaster layers with each other.}
}
\value{
A SpatRaster.
}
\description{
Loops over layers of SpatRasters within supplied list object and calculates one resulting layer per layer-group across SpatRasters in accordance with supplied FUN argument.
}
\examples{
Rast1 <- Download.KlimaiNorge2100(
    Variable = "Mean Air Temperature",
    DateStart = "1990-01-01",
    DateStop = "1990-12-31",
    Model = "CNRM_CCLM",
    Scenario = "rcp85",
    Cores = 1,
    Dir = getwd(),
    FileName = "CNRM_CCLM",
    Compression = NA,
    RemoveTemporary = TRUE
)

Rast2 <- Download.KlimaiNorge2100(
    Variable = "Mean Air Temperature",
    DateStart = "1990-01-01",
    DateStop = "1990-12-31",
    Model = "CNRM_RCA",
    Scenario = "rcp85",
    Cores = 1,
    Dir = getwd(),
    FileName = "CNRM_RCA",
    Compression = NA,
    RemoveTemporary = TRUE
)

Handle.ListLayers(Rast_ls = list(Rast1, Rast2), FUN = mean)

}
