% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.CDSFunctionality.r
\name{Helper.CDS.ExecuteRequests}
\alias{Helper.CDS.ExecuteRequests}
\title{Executing CDS Requests}
\usage{
Helper.CDS.ExecuteRequests(CDSRequests_ls, API_User, API_Key)
}
\arguments{
\item{CDSRequests_ls}{List. Each element holding a list object representing a CDS request. Created by \code{\link{Helper.CDS.MakeRequest}}.}

\item{API_User}{Character. CDS API User}

\item{API_Key}{Character. CDS API Key}
}
\value{
Character. A vector of FileNames to load in subsequent scripts.
}
\description{
Loops over a list of CDS requests and stages, then listens for them sequentially.
}
\seealso{
\code{\link{Helper.CDS.MakeRequest}}
}
