% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.CDSFunctionality.r
\name{Helper.CDS.RegisterAPICredentials}
\alias{Helper.CDS.RegisterAPICredentials}
\title{Register CDS API Credentials}
\usage{
Helper.CDS.RegisterAPICredentials(API_User, API_Key)
}
\arguments{
\item{API_User}{Character. CDS API User}

\item{API_Key}{Character. CDS API Key}
}
\value{
No R object. An addition to the keychain if necessary.
}
\description{
Just checks if provided API user and Key have already been added to keychain and adds them if necessary.
}
\seealso{
\code{\link{Helper.CDS.ExecuteRequests}}.
}
