% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.InputChecker.r
\name{Helper.InputChecker}
\alias{Helper.InputChecker}
\title{Check validity of user input}
\usage{
Helper.InputChecker(InCheck_ls)
}
\arguments{
\item{InCheck_ls}{}
}
\value{
Nothing. Only generates an error if a check did not succeed.
}
\description{
Loops over contents of a named list of lists where each sublist contains the pointers Input, Allowed, and Operator.
}
\examples{
Variable <- "Maximum Air Temperature"
DateStart <- "1971-01-01"
DateStop <- "1971-08-31"
Model <- "CNRM_CCLM"
Scenario <- "rcp85"

Start <- as.POSIXct(paste0(DateStart, "00:00:00"), tz = "UTC")
Stop <- as.POSIXct(paste0(DateStop, "00:00:00"), tz = "UTC")

InCheck_ls <- list(
    Variable = list(
        Input = Variable,
        Allowed = Meta.Variables("KlimaiNorge2100")$name,
        Operator = "in"
    ),
    Time = list(
        Input = c(Start, Stop),
        Allowed = Meta.QuickFacts("KlimaiNorge2100")$time$extent,
        Operator = "exceeds"
    ),
    Models = list(
        Input = Model,
        Allowed = Meta.QuickFacts("KlimaiNorge2100")$models,
        Operator = "in"
    ),
    Scenarios = list(
        Input = Scenario,
        Allowed = Meta.QuickFacts("KlimaiNorge2100")$scenarios,
        Operator = "in"
    )
)

Helper.InputChecker(InCheck_ls)
}
