% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.Progress.r
\name{Helper.Progress}
\alias{Helper.Progress}
\title{Create a progress bar to be used for progress tracking in package functions}
\usage{
Helper.Progress(IterLength, Text)
}
\arguments{
\item{IterLength}{Numeric. How many iterations the progress bar ought to track.}

\item{Text}{Character. What text to display as the action being performed in the progress bar.}
}
\value{
A progress progress bar object.
}
\description{
Creates a progress bar as defined by the progress package for use in other functions.
}
\examples{
Iter_n <- 100
pb <- Helper.Progress(IterLength = Iter_n, Text = "Elapsing")
for(i in 1:Iter_n){
    Sys.sleep(0.25)
    pb$tick(tokens = list(layer = i))
}

}
