% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.Threshold.r
\name{Helper.Threshold}
\alias{Helper.Threshold}
\title{Check validity of user input}
\usage{
Helper.Threshold(
  Raster,
  Operator,
  Threshold,
  ReturnValues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Raster}{A SpatRaster within which coverage should be identified}

\item{Operator}{Character. One of ">", "<", ">=", "<=", "==", "!=" to evaluate Raster cells against Threshold value.}

\item{Threshold}{Numeric or Character. Threshold to evaluate Raster cells against using Operator value.}

\item{ReturnValues}{Logical. Whether to return Raster of values matching thresholding (TRUE) or only logical raster indicating where thresholding is matched (FALSE).}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Loops over contents of a named list of lists where each sublist contains the pointers Input, Allowed, and Operator.
}
\examples{
Data_rast <- rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1:31]]
Helper.Threshold(
    Raster = Data_rast,
    Operator = "<",
    Threshold = 2650
)
Helper.Threshold(
    Raster = Data_rast,
    Operator = "<",
    Threshold = 2650,
    ReturnValues = TRUE
)

}
