% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_AccessCF.r
\name{Helper_AccessCF}
\alias{Helper_AccessCF}
\title{Execute direct download calls with ncdfCF}
\usage{
Helper_AccessCF(URLs, variable, subset = list(), verbose = TRUE)
}
\arguments{
\item{URLs}{Character. Vector of URLs for download.}

\item{variable}{Character. Variable name to extract from the netCDF files.}

\item{subset}{List, optional. A list with the ranges along the axes of the data variable to subset.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A CFVariable object containing the requested data.
}
\description{
Loops over all supplied URLs assuming these query netCDF files from open hosts. These are then registered and subsetted for variable and extent using ncdfCF before being downloaded and appended by the "time" axis.
}
\author{
Erik Kusch, Patrick Van Laake
}
