% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_AttrOverride.r
\name{Helper_AttrOverride}
\alias{Helper_AttrOverride}
\title{Overriding of attributes of one SpatRaster with those from another}
\usage{
Helper_AttrOverride(spatRasterFrom, spatRasterTo)
}
\arguments{
\item{spatRasterFrom}{A SpatRaster whose attributes (time, varnames, longnames, units, and metags) are to be assigned to another SpatRaster.}

\item{spatRasterTo}{A SpatRaster whose attributes (time, varnames, longnames, units, and metags) are to be overriden by those from another SpatRaster.}
}
\value{
A spatRaster.
}
\description{
Override attributes (time, varnames, longnames, units, and metags) of a SpatRaster with those from another SpatRaster.
}
\examples{
## loading example data
TX_rast <- NC_Read(system.file("extdata", "Sognefjord_TX.nc", package = "ClimHub"))[[1]] # a 1x1km spatRaster
RR_rast <- NC_Read(system.file("extdata", "Sognefjord_RR.nc", package = "ClimHub"))[[1]] # a 1x1km spatRaster
Helper_AttrOverride(spatRasterFrom = TX_rast, spatRasterTo = RR_rast)
}
\author{
Erik Kusch
}
