% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_CallString.r
\name{Helper_CallString}
\alias{Helper_CallString}
\title{Prepare metadata string for storing in NetCDF}
\usage{
Helper_CallString(callargs, functionName, citation = NULL)
}
\arguments{
\item{callargs}{Vector. A named vector of function call arguments to be stored as metadata.}

\item{functionName}{Character. The name of the function being called and whose arguments are stored in callargs.}

\item{citation}{Optional, Character. A citation string to be prepended to the metadata.}
}
\value{
A named vector suitable for storing as metadata attributes in NetCDF files.
}
\description{
Takes function call arguments and prepares a named vector suitable for storing as metadata attributes in NetCDF files. Optionally, also preprends a citation string.
}
\examples{
callargs <- c(variable = "TS", dateStart = "1961-08-01 00", dateStop = "1961-08-02 18")
functionName <- "Access_NORA3"
citation <- "Please cite NORA3 data as: ..."
Helper_CallString(callargs, functionName, citation)
}
\author{
Erik Kusch
}
