% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_CDSFunctionality.r
\name{Helper_CDS_ExecuteRequests}
\alias{Helper_CDS_ExecuteRequests}
\title{Execute CDS Requests}
\usage{
Helper_CDS_ExecuteRequests(cdsRequests, apiUser, apiKey)
}
\arguments{
\item{cdsRequests}{Optional, list. Each element holding a list object representing a CDS request. Created by \code{\link{Helper_CDS_MakeRequest}}.}

\item{apiUser}{Optional, character. CDS API User.}

\item{apiKey}{Optional, character. CDS API Key.}
}
\value{
Character. A vector of FileNames to load in subsequent scripts.
}
\description{
Loops over a list of CDS requests and stages, then listens for them sequentially.
}
\seealso{
\code{\link{Helper_CDS_MakeRequest}}
}
\author{
Erik Kusch
}
