% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_CDSFunctionality.r
\name{Helper_CDS_MakeRequests}
\alias{Helper_CDS_MakeRequests}
\title{Form CDS Requests}
\usage{
Helper_CDS_MakeRequests(queryTimeWindows, queryParams, dir = getwd())
}
\arguments{
\item{queryTimeWindows}{Optional, list. List of date ranges created by \code{\link{Helper_CDS_MakeRequestWindows}}.}

\item{queryParams}{Optional, list. Query parameters to be added to request.}

\item{dir}{Optional, character. Where to store CDS request outcomes.}
}
\value{
List. Each element holding a list object representing a CDS request.
}
\description{
Loops over time windows of defined size and creates a list of CDS requests.
}
\seealso{
\code{\link{Helper_CDS_MakeRequestWindows}}, \code{\link{Helper_CDS_RegisterAPICredentials}}, \code{\link{Helper_CDS_ExecuteRequests}}.

Make list of CDS Requests
Return list
}
\author{
Erik Kusch
}
