% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_CDSFunctionality.r
\name{Helper_CDS_MakeRequestWindows}
\alias{Helper_CDS_MakeRequestWindows}
\title{Create Time Windows for CDS Queries}
\usage{
Helper_CDS_MakeRequestWindows(allowedSeq, queriedSeq, tChunkSize)
}
\arguments{
\item{allowedSeq}{Optional, sequence. Sequence supported by queried data set.}

\item{queriedSeq}{Optional, sequence. Sequence queried by user.}

\item{tChunkSize}{Optional, numeric. Maximum amount of layers to include in each query.}
}
\value{
List. A list of dates for individual CDS queries as used by \code{\link{Helper_CDS_MakeRequest}}.
}
\description{
Make a list holding date ranges for which to make individual CDS queries.
}
\author{
Erik Kusch
}
