% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_CDSFunctionality.r
\name{Helper_CDS_RegisterAPICredentials}
\alias{Helper_CDS_RegisterAPICredentials}
\title{Register CDS API Credentials}
\usage{
Helper_CDS_RegisterAPICredentials(apiUser, apiKey)
}
\arguments{
\item{apiUser}{Optional, character. CDS API User.}

\item{apiKey}{Optional, character. CDS API Key.}
}
\value{
No R object. An addition to the keychain if necessary.
}
\description{
Just checks if provided API user and Key have already been added to keychain and adds them if necessary.
}
\seealso{
\code{\link{Helper_CDS_ExecuteRequests}}.
}
\author{
Erik Kusch
}
