% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_DailyBootstrap.r
\name{Helper_DailyBootstrap}
\alias{Helper_DailyBootstrap}
\title{Prepare Bootstrap Intervals of Daily Time Series across Years}
\usage{
Helper_DailyBootstrap(dates, bootstrapWindow = 5)
}
\arguments{
\item{dates}{A character vector of dates in the format "YYYY-MM-DDTHH:MM:SS" (e.g., "1961-01-01T00:00:00") representing the time series for which bootstrap intervals should be created. Same as is obtained with the \code{dimnames(ds[["time"]])} of a \code{CFDataset} object called \code{ds}.}

\item{bootstrapWindow}{Numeric, uneven integer. Length of bootstrapping interval around each target date.}
}
\value{
A list of character vectors, each containing the dates that fall within the bootstrap window around each target date. The names of the list correspond to the day of year (in "MM-DD" format) for which the bootstrap intervals are created. Interval creation assumes a leap year (i.e., 366 days) to ensure that the same day of year is always included in the bootstrap intervals, even for February 29th.
}
\description{
Create bootstrap intervals of daily time series across years for a given time sequence. This function is needed for baseperiod quantile calculations for \code{Metrics_ETCCDI}.
}
\examples{
time_series <- seq(
    as.POSIXct("1961-01-01 00:00:00", tz = "UTC"),
    as.POSIXct("1999-12-31 23:00:00", tz = "UTC"),
    by = "1 day"
)
Time_seq <- format(time_series, "\%Y-\%m-\%dT\%H:\%M:\%S")
Helper_DailyBootstrap(dates = Time_seq)

}
\author{
Erik Kusch
}
