% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_DirectDownload.r
\name{Helper_DirectDownload}
\alias{Helper_DirectDownload}
\title{Execute direct download calls}
\usage{
Helper_DirectDownload(url, fileName, cores, verbose = TRUE)
}
\arguments{
\item{url}{Character. Vector of URLs for download.}

\item{fileName}{Character. Vector of fileName for downloaded files.}

\item{cores}{Integer. Number of cores for parallelisation if desired.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A vector of filenames. Same as the fileName argument.
}
\description{
Loops over all supplied file url and file fileName, assuming these query NetCDF file from open hosts and downloads these files. In the process, it checks whether files that have been downloaded can be loaded from the disk. Parallelisation can speed this process up a lot.
}
\examples{
Helper_DirectDownload(
    url = c(
        "https://thredds.met.no/thredds/fileServer/nora3/1961/08/01/00/fc1961080100_003_sfx.nc",
        "https://thredds.met.no/thredds/fileServer/nora3/1961/08/01/06/fc1961080106_003_sfx.nc"
    ),
    fileName = c("TEMP_fc1961080100_003_sfx.nc", "TEMP_fc1961080106_003_sfx.nc"),
    cores = 2
)
}
\author{
Erik Kusch
}
