% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_ETCCDIGSL.r
\name{Helper_ETCCDIGSL}
\alias{Helper_ETCCDIGSL}
\title{Calculate Growing Season Length (GSL)}
\usage{
Helper_ETCCDIGSL(TM)
}
\arguments{
\item{TM}{A CFVariable containing daily mean temperatures (or similar) to be used for GSL calculation.}
}
\value{
A CFVariable.
}
\description{
Calculates Growing Season Length as defined by ETCCDI with different treatment of northern and southern hemisphere. For the northern hemisphere, GSL is defined as the number of days between the first occurrence of 6 consecutive days with mean temperature above 5°C after Jan 1st and the first occurrence of 6 consecutive days with mean temperature below 5°C after Jul 1st. For the southern hemisphere, GSL is defined as the number of days between the first occurrence of 6 consecutive days with mean temperature above 5°C after Jul 1st and the first occurrence of 6 consecutive days with mean temperature below 5°C after Jan 1st of the following year. If no valid start or end date is found, GSL is set to NA for that year and grid cell.
}
\examples{
\dontrun{
TX <- NC_Read("inst/extdata/KiN_tasmax_2050.nc")[["tasmax"]]
TN <- NC_Read("inst/extdata/KiN_tasmin_2050.nc")[["tasmin"]]
TM <- (TN + TX) / 2
GSL <- Helper_ETCCDIGSL(TM)
}
}
\author{
Erik Kusch
}
