% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_EvalLoopText.r
\name{Helper_EvalLoopText}
\alias{Helper_EvalLoopText}
\title{Loop over a set numnber of iterations of a string to harmonise looping in package}
\usage{
Helper_EvalLoopText(
  loopText,
  iters,
  objects,
  packages = "",
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{loopText}{Character. String to be valuated in Loop.}

\item{iters}{Numeric. Sequence of iterator values for "Iter" object to take.}

\item{objects}{List. List of objects the evaluator needs access to.}

\item{packages}{Character. Vector containing R package names which are needed for loopText evaluation.}

\item{cores}{Integer. Number of cores for parallelisation if desired.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A list of whatever the loopText returns per iteration
}
\description{
Iteraters over a set number of iterations using a loop-text string which gets evaluated at each iteration.
}
\examples{
## objects which to use in the loops
Vec1 <- Vec2 <- 1:100
## loopText to be evaluated
loopText <- "Sys.sleep(0.01); Vec1[Iter] + Vec2[Iter]"

## Single-Core Exection
pb <- Helper_Progress(iterLength = length(Vec1), text = "TestLoop")
Helper_EvalLoopText(loopText = loopText, 
    iters = 1:length(Vec1), 
    objects = list("Vec1" = Vec1, "Vec2" = Vec2, "pb" = pb)
    )

## Parallel Exection with progress
pb <- Helper_Progress(iterLength = length(Vec1), text = "TestLoop")
Helper_EvalLoopText(loopText = loopText, 
    iters = 1:length(Vec1), 
    objects = list("Vec1" = Vec1, "Vec2" = Vec2, "pb" = pb),
    cores = 2
    )

## Parallel Exection without progress
pb <- Helper_Progress(iterLength = length(Vec1), text = "TestLoop")
Helper_EvalLoopText(loopText = loopText, 
    iters = 1:length(Vec1), 
    objects = list("Vec1" = Vec1, "Vec2" = Vec2, "pb" = pb),
    cores = 2,
    verbose = FALSE
    )
}
\author{
Erik Kusch
}
