% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_FileCheck.r
\name{Helper_FileCheck}
\alias{Helper_FileCheck}
\title{Checking if a file already exists}
\usage{
Helper_FileCheck(fileName, loadFun, load = TRUE, verbose = TRUE)
}
\arguments{
\item{fileName}{Filename including directory}

\item{loadFun}{function with which to load filetype of fileName}

\item{load}{Logical. Whether to load the data or not}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}
}
\value{
Either a data object or NULL
}
\description{
If a file already exists in a given place, load that file.
}
\examples{
## returning file contents as R object if file present and prompted to load
Helper_FileCheck(fileName = system.file("extdata", "KiN_rast.nc", package = "ClimHub"), loadFun = NC_Read, load = TRUE, verbose = TRUE)
## returning NULL if file does not exist
Helper_FileCheck(fileName = "", loadFun = NC_Read, load = TRUE, verbose = TRUE)
}
\author{
Erik Kusch
}
