% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_LoadFiles.r
\name{Helper_LoadFiles}
\alias{Helper_LoadFiles}
\title{Load list of NetCDF files and stack them into a SpatRaster}
\usage{
Helper_LoadFiles(fileName, dates = NULL, verbose = TRUE)
}
\arguments{
\item{fileName}{Character. Vector of absolute filenames to be loaded.}

\item{dates}{POSIXct. Optional vector of times to apply to list-loaded data}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A spatraster object.
}
\description{
Iterates of a number of filenames, pointing to NetCDFs, and loads them into one SpatRaster.
}
\examples{
Helper_LoadFiles(
   fileName = c(
       system.file("extdata", "KiN_rast.nc", package = "ClimHub"), 
       system.file("extdata", "KiN_rast.nc", package = "ClimHub")
       )
   )
}
\author{
Erik Kusch
}
