% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_MakeUTC.r
\name{Helper_MakeUTC}
\alias{Helper_MakeUTC}
\title{Resolve time zones as requested by user and UTC format with which to query from CDS}
\usage{
Helper_MakeUTC(dates = NULL)
}
\arguments{
\item{dates}{A vector of POSIXct objects}
}
\value{
A data frame on input dates respective to user-queried timezone and their UTC counterparts.
}
\description{
Create UTC counterparts of user-input dates for CDS queries
}
\examples{
IN_DateStart <- as.POSIXct("1995-01-01 00:00", tz = "CET")
IN_DateStop <- as.POSIXct("2005-01-01 23:00", tz = "CET")
Dates_df <- Helper_MakeUTC(dates = c(IN_DateStart, IN_DateStop))
Dates_df
}
\author{
Erik Kusch
}
