% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_PolySummary.r
\name{Helper_PolySummary}
\alias{Helper_PolySummary}
\title{Summarise Cell Values of SpatRaster by Cell Polygon Structure of Another}
\usage{
Helper_PolySummary(spatRasterFrom, spatRasterTo, fun = "mean", ...)
}
\arguments{
\item{spatRasterFrom}{A SpatRaster whose cells are to be summarised.}

\item{spatRasterTo}{A SpatRaster to whose cell boundaries/structure values in cells of spatRasterFrom should be summarised.}

\item{fun}{User-defined function by which to summarise values of cells. Supported functions are mean, sum, min, max and table.}

\item{...}{additional arguments passed to \code{fun}, such as na.rm=TRUE}
}
\value{
A spatRaster.
}
\description{
Convert to cell structure of one SpatRaster into polygons (a SpatVector) and extract values for these from another SpatRaster resulting in a SpatRaster with the cell structure of the former but values from the latter.
}
\examples{
## loading example data and making a coarse raster
Fine_rast <- NC_Read(system.file("extdata", "Sognefjord_TX.nc", package = "ClimHub"))[[1]] # a 1x1km spatRaster
# Viz_SpatRast(Fine_rast)
Coarse_rast <- terra::aggregate(Fine_rast, fact = 3)
# Viz_SpatRast(Coarse_rast)
## aggregation
Aggr_rast <- Helper_PolySummary(spatRasterFrom = Fine_rast, spatRasterTo = Coarse_rast, na.rm = TRUE)
# Viz_SpatRast(Aggr_rast)
## disaggregation
Disagg_rast <- Helper_PolySummary(spatRasterFrom = Coarse_rast, spatRasterTo = Fine_rast, na.rm = TRUE)
# Viz_SpatRast(Disagg_rast)
}
\author{
Erik Kusch
}
