% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Progress.r
\name{Helper_Progress}
\alias{Helper_Progress}
\title{Create a progress bar to be used for progress tracking in package functions}
\usage{
Helper_Progress(iterLength, text)
}
\arguments{
\item{iterLength}{Numeric. How many iterations the progress bar ought to track.}

\item{text}{Character. What text to display as the action being performed in the progress bar.}
}
\value{
A progress progress bar object.
}
\description{
Creates a progress bar as defined by the progress package for use in other functions.
}
\examples{
Iter_n <- 100
pb <- Helper_Progress(iterLength = Iter_n, text = "Elapsing")
for(i in 1:Iter_n){
    Sys.sleep(0.25)
    pb$tick(tokens = list(layer = i))
}
}
\author{
Erik Kusch
}
