% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Threshold.r
\name{Helper_Threshold}
\alias{Helper_Threshold}
\title{Apply tresholding to CFVariable contents}
\usage{
Helper_Threshold(
  CFVariable,
  operator,
  threshold,
  threshMode = NULL,
  returnValues = FALSE,
  returnSummary,
  returnTResolution = "year"
)
}
\arguments{
\item{CFVariable}{A CFVariable within which thresholding should be applied}

\item{operator}{Character. One of ">", "<", ">=", "<=", "==", "!=" to evaluate CFVariable cells against threshold value.}

\item{threshold}{Numeric or Character. threshold to evaluate CFVariable cells against using operator value.}

\item{threshMode}{Character, optional. If supplying a CFVariable containing baseline quantiles for ETCCDI calculation, setting \code{threshMode = "ETCCDIQuantiles"} will internally apply location and day-of-year specific thresholds to \code{CFVariable} input. Default is \code{NULL} avoiding this behaviour.}

\item{returnValues}{Logical. Whether to return CFVariable of values matching thresholding (TRUE) or only logical raster indicating where thresholding is matched (FALSE).}

\item{returnSummary}{Function. Summary function for tresholded CFVariable.}

\item{returnTResolution}{Character. Temporal resolution of summary of tresholded CFVariable.}
}
\value{
A CFVariable.
}
\description{
Evaluates threshold criteria defined by user against a CFVariable and reports summaries through the \verb{$summarise} method of \code{ncdfCF}.
}
\examples{
Data_CF <- NC_Read("inst/extdata/KiN_tas.nc")[["tas"]]
Helper_Threshold(
    CFVariable = Data_CF,
    operator = "<",
    threshold = 273.15
)
Helper_Threshold(
    CFVariable = Data_CF,
    operator = "<",
    threshold = 273.15,
    returnValues = TRUE,
    returnSummary = mean
)
}
\author{
Erik Kusch
}
