% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_TimeSplit.r
\name{Helper_TimeSplit}
\alias{Helper_TimeSplit}
\title{Split raster layers by time component}
\usage{
Helper_TimeSplit(spatRaster, tResolution)
}
\arguments{
\item{spatRaster}{A SpatRaster with a time dimension/component.}

\item{tResolution}{Character. Indicating by which division of the time component to split. Supported values are c("Year", "Season", "Month", "YearMonth")}
}
\value{
A list of SpatRasters.
}
\description{
Creates a list object containing layers of input SpatRaster object according to time component and user-defined split.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_AT.nc", package = "ClimHub"))[[1:360]]
Helper_TimeSplit(spatRaster = Data_rast, tResolution = "Year")
Helper_TimeSplit(spatRaster = Data_rast, tResolution = "Season")
Helper_TimeSplit(spatRaster = Data_rast, tResolution = "Month")
Helper_TimeSplit(spatRaster = Data_rast, tResolution = "YearMonth")
}
\author{
Erik Kusch
}
