% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_VarDFtoJSON.r
\name{Helper_VarDFtoJSON}
\alias{Helper_VarDFtoJSON}
\title{Write Data Frame to JSON}
\usage{
Helper_VarDFtoJSON(df, outfile = "variables.json")
}
\arguments{
\item{df}{A data frame to be written to JSON.}

\item{outfile}{Character string with the path to the output JSON file.}
}
\value{
A JSON file saved to the specified path.
}
\description{
Writes a data frame to a JSON file. Designed to simplify creation of JSON files from variable metadata data frames.
}
\examples{
NORA3_df <- Discovery_Variables(dataSet = "NORA3")
Helper_VarDFtoJSON(NORA3_df, "NORA3_variables.json")

}
\author{
Erik Kusch
}
