% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Jotunheimen_sf}
\alias{Jotunheimen_sf}
\title{Shapefile of boundaries of Jotunheimen national park}
\format{
a sf POLYGON
}
\usage{
Jotunheimen_sf
}
\description{
An sf object containing a polygon describing the boundaries of Jotunheimen national park
}
\keyword{datasets}
