% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.Citation.r
\name{Meta.Citation}
\alias{Meta.Citation}
\title{DOI of data set}
\usage{
Meta.Citation(dataset = "NULL")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta\.Library()}}.}
}
\value{
Character. Citation string for data set.
}
\description{
Read and return DOI of data set for easy citation.
}
\examples{
Meta.Citation(dataset = "NORA3")

}
\seealso{
\code{\link{Meta\.Library}}, \code{\link{Meta.Read}}, \code{\link{Meta.Variables}}, \code{\link{Meta.QuickFacts}}, \code{\link{Meta.DOI}.
}
