% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.DataSet.Check.r
\name{Meta.DataSet.Check}
\alias{Meta.DataSet.Check}
\title{Check dataset string validity}
\usage{
Meta.DataSet.Check(dataset)
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.Library()}}.}
}
\value{
Throws an error message if user-provided dataset string is not recognised in metadata library.
}
\description{
Check if dataset string is resolvable in metadata library.
}
\examples{
Meta.DataSet.Check(dataset = "NORA3")
# Meta.DataSet.Check(dataset = "NULL")
}
\seealso{
\code{\link{Meta.Library}}, \code{\link{Meta.Variables}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}.
}
