% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.Library.r
\name{Meta.Library}
\alias{Meta.Library}
\title{List out all supported data sets}
\usage{
Meta.Library()
}
\value{
A vector of supported datasets.
}
\description{
Provide an overview of all data sets for which metadata files are present.
}
\examples{
Meta.Library()

}
\seealso{
\code{\link{Meta.Read}}, \code{\link{Meta.Variables}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}, \code{\link{Meta.Citation}}.
}
