% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.Read.r
\name{Meta.Read}
\alias{Meta.Read}
\title{Data set overview}
\usage{
Meta.Read(dataset = "NULL")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.Library()}}.}
}
\value{
List. Contains information of data set, type, variables, resolution, citation, etc.
}
\description{
Read and return metadata for specific data set.
}
\examples{
Meta.Read(dataset = "NORA3")

}
\seealso{
\code{\link{Meta.Library}}, \code{\link{Meta.Variables}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}, \code{\link{Meta.Citation}}.
}
