% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.Variables.r
\name{Meta.Variables}
\alias{Meta.Variables}
\title{Variables available within data set}
\usage{
Meta.Variables(dataset = "NULL")
}
\arguments{
\item{dataset}{Character. Name of data set. Usually a set of words separated by dashes. See possible datasets by calling \code{\link{Meta.Library()}}.}
}
\value{
Data frame. Contains at least 3 columns:
\itemize{
\item{name}{Variable clear name and what is used to query a download}
\item{unit}{Unit of measurement}
\item{issues}{potential known issues with this variable}
}
May contain additional columns mostly used for backend operations.
}
\description{
Read and return overview of variables available for specific data set.
}
\examples{
Meta.Variables(dataset = "NORA3")

}
\seealso{
\code{\link{Meta.Library}}, \code{\link{Meta.Read}}, \code{\link{Meta.DOI}}, \code{\link{Meta.QuickFacts}}, \code{\link{Meta.Citation}}.
}
