% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics.ETCCDI.r
\name{Metrics.ETCCDI}
\alias{Metrics.ETCCDI}
\title{CALCULATE EXPERT TEAM CLIMATE CHANGE DETECTION INDICES (ETCCDI) METRICS ======================================================
Calculate ETCCDI indices from list input}
\usage{
Metrics.ETCCDI(Rasters)
}
\arguments{
\item{Rasters}{List of SpatRasters. Names of elements bust be "TX" and "TN", holding maximum and minimum daily air temperature respectively.}
}
\value{
A named list of SpatRasters with each element corresponding to an ETCCDI.
}
\description{
This function calculates \href{https://etccdi.pacificclimate.org/list_27_indices.shtml}{ETCCDIs} from a named list of SpatRaster objects. Currently, only the first four ETCCDI are supported.
}
\examples{
\dontrun{
TX_rast <- Download.KlimaiNorge2100(
    Variable = "Maximum Air Temperature",
    DateStart = "1995-01-01",
    DateStop = "1996-12-31",
    Model = "CNRM_CCLM",
    FileName = "KiN_TX",
)

TN_rast <- Download.KlimaiNorge2100(
    Variable = "Minimum Air Temperature",
    DateStart = "1995-01-01",
    DateStop = "1996-12-31",
    Model = "CNRM_CCLM",
    FileName = "KiN_TN",
)

Metrics.ETCCDI(Rasters = list(TX = TX_rast, TN = TN_rast))
}

}
