% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Read.r
\name{NC_Read}
\alias{NC_Read}
\title{Read a netCDF file}
\usage{
NC_Read(resource)
}
\arguments{
\item{resource}{Character. Fully qualified file name for a netCDF file on a
local file system or the URL to a netCDF resource on a THREDDS server.}
}
\value{
A \code{ncdfCF} data set instance.
}
\description{
Open and read a netCDF file including its metadata. Data is
returned as a \code{ncdfCF} object.
}
\examples{
ds <- NC_Read(resource = system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
}
\author{
Patrick Van Laake
}
