% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Write.r
\name{NC_Write}
\alias{NC_Write}
\title{Write NetCDF files and their metadata with a progress bar for netcdf layers}
\usage{
NC_Write(
  spatRaster,
  fileName,
  compression = NA,
  varName,
  longName,
  unit,
  meta = NULL,
  verbose = TRUE
)
}
\arguments{
\item{spatRaster}{SpatRaster}

\item{fileName}{Filename including directory}

\item{compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}

\item{varName}{Character. Variable name to be saved as varname in NC output.}

\item{longName}{Character. Long varName name.}

\item{unit}{Character. Unit in which Variable (varName) is recorded.}

\item{meta}{Named vector of metadata attributes}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster with metadata written to the disk
}
\description{
Write NetCDF files and their metadata with a progress bar for netcdf layers. Adapted from the .writeCDF method in the terra package: https://github.com/rspatial/terra/blob/master/R/ncdf.R
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
Data_rast <- terra::crop(Data_rast, c(0, 7e4, 6.7e6, 6.77e6))
NC_Write(spatRaster = Data_rast, 
        fileName = "temp.nc", 
        compression = NA, 
        varName = terra::varnames(Data_rast), 
        longName = "testvar", 
        unit = "K", 
        meta = c("Citation" = "ClimHub Helper_WriteNC Test")
    )
unlink("temp.nc")
}
\author{
Erik Kusch
}
