% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Nor2K_sf}
\alias{Nor2K_sf}
\title{SpatialFeatures Object of select summits Norway}
\format{
a sf POINT collection
}
\usage{
Nor2K_sf
}
\description{
Point data of summits throuhout Norway with an elevation greater than 2,000m asl. Source for coordinates and elevation: https://en.wikipedia.org/wiki/List_of_mountains_in_Norway_by_height.
}
\keyword{datasets}
