% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ListLayers.r
\name{Process_ListLayers}
\alias{Process_ListLayers}
\title{Layer-wise application on SpatRasters in a list}
\usage{
Process_ListLayers(ls, fun = mean, verbose = TRUE)
}
\arguments{
\item{ls}{List. List of SpatRasters whose layers to compute with each other. All SpatRasters need to have the same layer count.}

\item{fun}{Function. Function by which to summarise/compute SpatRaster layers with each other.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Loops over layers of SpatRasters within supplied list object and calculates one resulting layer per layer-group across SpatRasters in accordance with supplied fun argument.
}
\examples{
## load the first 2 days of the KiN temperature data provided with ClimHub to two separate list slots and articifially increase values by 1 for the second list item
ls <- list(
    Base = terra::rast(system.file("extdata", "KiN_AT.nc", package = "ClimHub"))[[1:2]],
    Base1 = terra::rast(system.file("extdata", "KiN_AT.nc", package = "ClimHub"))[[1:2]]+1
    )
## calculate difference between each layer in the two list slots
Diff_rast <- Process_ListLayers(ls = ls, fun = diff)
Diff_rast # all values in this 2-layer raster, ought to be 1 or NA

}
\author{
Erik Kusch
}
