% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.Buffer.r
\name{Spatial.Buffer}
\alias{Spatial.Buffer}
\title{Square Buffers Around Point Data}
\usage{
Spatial.Buffer(SF, Buffer = 1000)
}
\arguments{
\item{SF}{An sf POINT object}

\item{Buffer}{Size of buffer (unit dependant on projection of input SF).}
}
\value{
An sf polygon made up of individual square buffers around point-location input.
}
\description{
Allow for drawing of buffer zones around point-location data for downloading and kriging of spatial data around point-locations. Overlapping individual buffers are merged.
}
\examples{
data(Nor2K_sf)
Spatial.Buffer(Nor2K_sf, Buffer = 1e3) # 1km buffer
}
