% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.Extract.r
\name{Spatial.Extract}
\alias{Spatial.Extract}
\title{Extracting data from SpatRaster with a SpatialFeaturesObject}
\usage{
Spatial.Extract(Raster, SF, FUN = mean)
}
\arguments{
\item{Raster}{A SpatRaster within which coverage should be identified}

\item{SF}{Either an sf polygon(-collection) or an sf point(-collection).}

\item{FUN}{User-defined function by which to aggregate values of cells within a polygon. Supported functions are mean, sum, min, max and table.}
}
\value{
A data.frame.
}
\description{
Extract data from SpatRaster corresponding to spatial features in SF object.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
data(Jotunheimen_sf)
Spatial.Extract(
    Raster = Data_rast,
    SF = Jotunheimen_sf,
    FUN = mean
)
data(Nor2K_sf)
Spatial.Extract(
    Raster = Data_rast,
    SF = Nor2K_sf
)
}
