% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.MakePoints.r
\name{Spatial.MakePoints}
\alias{Spatial.MakePoints}
\title{Transform data frame-type inputs into sf}
\usage{
Spatial.MakePoints(DataFrame, CRS = 4326)
}
\arguments{
\item{DataFrame}{A data.frame containing geo-referenced points with Lat and Lon columns}

\item{CRS}{Numeric. EPSG CRS number. See \href{https://epsg.io/}{here} for more information}
}
\value{
An sf POINT object.
}
\description{
Transform data frame with ID for querying functionality around point-loactions to SpatialPoints
}
