% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.Reproject.r
\name{Spatial.Reproject}
\alias{Spatial.Reproject}
\title{Reprojecting spatial data.}
\usage{
Spatial.Reproject(ProjFrom, ProjTo)
}
\arguments{
\item{ProjFrom}{Either a SpatRaster or SF object which is to be reprojected.}

\item{ProjTo}{Prefably, a character string identifying desired EPSG ID. Can also be a SpatRaster or SF object to whose projection the other opbject is to be reprojected. The latter specifications are less stable.}
}
\value{
A SpatRaster or SF object reprojected to the same CRS as in ProjTo.
}
\description{
Reprojecting one spatial data product to match another.
}
