% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Aggregate.r
\name{Spatial_Aggregate}
\alias{Spatial_Aggregate}
\title{Spatial Aggregation of SpatRaster Data}
\usage{
Spatial_Aggregate(
  spatRasterFrom,
  spatRasterTo,
  fun = "mean",
  ...,
  factor,
  cores = 1
)
}
\arguments{
\item{spatRasterFrom}{A SpatRaster whose cells are to be aggregated.}

\item{spatRasterTo}{A SpatRaster to whose cell boundaries values in cells of spatRasterFrom should be aggregated.}

\item{fun}{User-defined function by which to aggregate values of cells (e.g., mean, sd, sum, min, max).}

\item{...}{additional arguments passed to \code{fun}, such as \code{na.rm=TRUE}}

\item{factor}{Integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). Passed on to \code{terra::aggregate}.}

\item{cores}{Integer. Number of cores for parallelisation if desired.}
}
\value{
A spatRaster.
}
\description{
Aggregation of spatraster data to either (1) match a pre-established spatraster or (2) coarsen spatial resolution by a specific factor.
}
\examples{
## loading example data
Fine_rast <- NC_Read(system.file("extdata", "Sognefjord_TX.nc", package = "ClimHub"))[[1]] # a 1x1km spatRaster
# Viz_SpatRast(Fine_rast)

## coarsening by a factor of 3
Coarse_rast1 <- Spatial_Aggregate(spatRasterFrom = Fine_rast, fun = "mean", factor = 3)
# Viz_SpatRast(Coarse_rast1)

## coarsening by a factor of 5 and extracting SD
Coarse_rast2 <- Spatial_Aggregate(spatRasterFrom = Fine_rast, fun = "sd", factor = 5)
# Viz_SpatRast(Coarse_rast2)

## coarsening by a factor of 3, but ignoring NAs
Coarse_rast3 <- Spatial_Aggregate(spatRasterFrom = Fine_rast, fun = "mean", na.rm = TRUE, factor = 3)
# Viz_SpatRast(Coarse_rast3)

## coarsening to match existing raster; takes longer but is more versatile
Coarse_rast4 <- Spatial_Aggregate(spatRasterFrom = Fine_rast, spatRasterTo = Coarse_rast3, fun = "mean", na.rm = TRUE)
# Viz_SpatRast(Coarse_rast4)
}
\seealso{
\code{\link{Spatial_Disaggregate}}
}
\author{
Erik Kusch
}
