% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Buffer.r
\name{Spatial_Buffer}
\alias{Spatial_Buffer}
\title{Square Buffers Around Point Data}
\usage{
Spatial_Buffer(sf, buffer = 1000)
}
\arguments{
\item{sf}{An sf POINT object}

\item{buffer}{Size of buffer (unit dependant on projection of input SF).}
}
\value{
An sf polygon made up of individual square buffers around point-location input.
}
\description{
Allow for drawing of buffer zones around point-location data for downloading and kriging of spatial data around point-locations. Overlapping individual buffers are merged.
}
\examples{
data(Nor2K_sf)
Spatial.Buffer(Nor2K_sf, Buffer = 1e3) # 1km buffer
}
\author{
Erik Kusch
}
