% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Disaggregate.r
\name{Spatial_Disaggregate}
\alias{Spatial_Disaggregate}
\title{Spatial Disaggregation of SpatRaster Data}
\usage{
Spatial_Disaggregate(spatRasterFrom, spatRasterTo, fun, ..., factor)
}
\arguments{
\item{spatRasterFrom}{A SpatRaster whose cells are to be disaggregated.}

\item{spatRasterTo}{A SpatRaster to whose cell boundaries values in cells of spatRasterFrom should be disaggregated.}

\item{fun}{User-defined function by which to disaggregate values of cells. Supported functions are (1) mean, sum, min, max, etc. when specifying \code{spatRasterTo} and (2) either "near" for nearest or "bilinear" for bilinear interpolation when specufing \code{factor} (passed on to \code{terra::disagg} function as \code{method} argument).}

\item{...}{additional arguments passed to \code{fun}, such as \code{na.rm=TRUE} when specifying \code{spatRasterTo}}

\item{factor}{Integer. Disaggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also disaggregating over layers). Passed on to \code{terra::disagg}.}
}
\value{
A spatRaster.
}
\description{
Disaggregation of spatraster data to either (1) match a pre-established spatraster or (2) coarsen spatial resolution by a specific factor.
}
\examples{
## loading example data
Coarse_rast <- NC_Read(system.file("extdata", "Sognefjord_TX.nc", package = "ClimHub"))[[1]] # a 1x1km spatRaster
# Viz_SpatRast(Coarse_rast)

## coarsening by a factor of 2
Fine_rast1 <- Spatial_Disaggregate(spatRasterFrom = Coarse_rast, fun = "near", factor = 2)
# Viz_SpatRast(Fine_rast1)

## coarsening by a factor of 5 and using bilinear method
Fine_rast2 <- Spatial_Disaggregate(spatRasterFrom = Coarse_rast, fun = "bilinear", factor = 5)
# Viz_SpatRast(Fine_rast2)

## coarsening to match existing raster; takes longer but is more versatile
\dontrun{
Fine_rast3 <- Spatial_Disaggregate(spatRasterFrom = Coarse_rast, spatRasterTo = Fine_rast1, fun = "mean", na.rm = TRUE)
# Viz_SpatRast(Fine_rast3)
}
}
\seealso{
\code{\link{Spatial_Aggregate}}
}
\author{
Erik Kusch
}
