% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Limit.r
\name{Spatial_Limit}
\alias{Spatial_Limit}
\title{Cropping & Range Masking with Edge Support}
\usage{
Spatial_Limit(spatRaster, shape, verbose = TRUE)
}
\arguments{
\item{spatRaster}{A SpatRaster within which coverage should be identified}

\item{shape}{Either a SpatExtent or an sf polygon(-collection) whose coverage of the raster object is to be found.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Cropped and masking the original SpatRaster (\code{spatRaster}) using supplied SpatExtent or shapefile (\code{shape}) and retaining all pixels which are even just partially covered.
}
\examples{
# single-layer exaples
## limit by sf object
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
data(Jotunheimen_sf)
# Kin_Limited <- Spatial_Limit(KiN_rast, Jotunheimen_sf) # this would throw an error due to different projections of the spatRaster and the sf object
KiN_rast <- Spatial_Reproject(KiN_rast, Jotunheimen_sf)
KiN_Limited <- Spatial_Limit(spatRaster = KiN_rast, shape = Jotunheimen_sf)
# Viz_SpatRast(KiN_Limited)

## limit by SpatExtent
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
KiN_Limited <- Spatial_Limit(spatRaster = KiN_rast, shape = terra::ext(-7e4, 3e5, 6.5e6, 7e6))
# Viz_SpatRast(KiN_Limited)

# multi-layer exaple
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
KiN_Limited <- Spatial_Limit(spatRaster = KiN_rast, shape = terra::ext(-7e4, 3e5, 6.5e6, 7e6))
# Viz_SpatRast(KiN_Limited, panelColumns = 5)
}
\seealso{
\code{\link{Spatial_Reproject}}.
}
\author{
Erik Kusch
}
