% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_MakePoints.r
\name{Spatial_MakePoints}
\alias{Spatial_MakePoints}
\title{Transform data frame-type inputs into sf}
\usage{
Spatial_MakePoints(dataFrame, crs = 4326)
}
\arguments{
\item{dataFrame}{A data.frame containing geo-referenced points with Lat and Lon columns}

\item{crs}{Numeric. EPSG CRS number. See \href{https://epsg.io/}{here} for more information}
}
\value{
An sf POINT object.
}
\description{
Transform data frame with ID for querying functionality around point-loactions to SpatialPoints
}
\examples{
data(Nor2K_sf)
df <- cbind(sf::st_drop_geometry(Nor2K_sf), sf::st_coordinates(Nor2K_sf))
colnames(df)[5:6] <- c("Lon", "Lat")
Spatial_MakePoints(df, sf::st_crs(Nor2K_sf))
}
\author{
Erik Kusch
}
