% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Reproject.r
\name{Spatial_Reproject}
\alias{Spatial_Reproject}
\title{Reprojecting spatial data.}
\usage{
Spatial_Reproject(projFrom, projTo, rasterResample = TRUE)
}
\arguments{
\item{projFrom}{Either a SpatRaster or SF object which is to be reprojected.}

\item{projTo}{Prefably, a character string identifying desired EPSG ID. Can also be a SpatRaster or SF object to whose projection the other opbject is to be reprojected. The latter specifications are less stable.}

\item{rasterResample}{Logical. If TRUE and both objects are SpatRasters, projFrom is also resampled to match the resolution and origin of ProjTo.}
}
\value{
A SpatRaster or SF object reprojected to the same CRS as in ProjTo.
}
\description{
Reprojecting one spatial data product to match another.
}
\examples{
## reprojecting raster to match specific EPSG
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]] # this data is stored in K*10
SpatRast_repro <- Spatial_Reproject(KiN_rast, 3995)
# Viz_SpatRast(SpatRast_repro)

## reprojecting raster to match sf
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
data(Jotunheimen_sf)
SpatRast_repro <- Spatial_Reproject(KiN_rast, Jotunheimen_sf)
# Viz_SpatRast(SpatRast_repro, sfObj = Jotunheimen_sf, sfColour = "white")

## reprojecting raster to match raster
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
Nora3_rast <- terra::rast(system.file("extdata", "Nora3_rast.nc", package = "ClimHub"))[[1]]
SpatRast_repro <- Spatial_Reproject(KiN_rast, Nora3_rast)
# Viz_SpatRast(SpatRast_repro)

## reprojecting sf to match raster
KiN_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
data(Nor2K_sf)
Nor2K_repro <- Spatial_Reproject(Nor2K_sf, KiN_rast)
# Viz_SpatRast(KiN_rast, sfObj = Nor2K_repro, sfColour = "white", sfFill = "white", sfSize = 2)
}
\author{
Erik Kusch
}
