% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.Aggregration.r
\name{Temporal.Aggregration}
\alias{Temporal.Aggregration}
\title{Carry out temporal aggregation}
\usage{
Temporal.Aggregration(Raster, TResolution, TStep, FUN)
}
\arguments{
\item{Raster}{A SpatRaster within which coverage should be identified}

\item{TResolution}{Character. User-specified temporal resolution}

\item{TStep}{Numeric. User-specified time step}

\item{FUN}{User-defined aggregation function}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal aggregation and carries it out.
}
