% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.Decumulation.r
\name{Temporal.Decumulation}
\alias{Temporal.Decumulation}
\title{Carry out decumulation of cumulatively stored data}
\usage{
Temporal.Decumulation(Raster, Interval, Mode, Cores = 1)
}
\arguments{
\item{Raster}{A SpatRaster within which coverage should be identified}

\item{Interval}{Number of layers in the raster belonging to the same interval of cumulation.}

\item{Mode}{Character. Mode of cumulative storage of values. Currently supported: CDS (first layer per day is the cumulative sum of the previous day).}

\item{Cores}{Integer. Number of cores for parallelisation if desired.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal cumulation and reverses the cumulative measurements into indvidual ones. Currently conceptualised to decumulate hourly CDS records.
}
