% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal_Aggregation.r
\name{Temporal_Aggregration}
\alias{Temporal_Aggregration}
\title{Carry out temporal aggregation}
\usage{
Temporal_Aggregration(spatRaster, tResolution, tStep, fun, verbose = TRUE)
}
\arguments{
\item{spatRaster}{A SpatRaster within which coverage should be identified}

\item{tResolution}{Character. User-specified temporal resolution}

\item{tStep}{Numeric. User-specified time step}

\item{fun}{User-defined aggregation function}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal aggregation and carries it out.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
Temporal_Aggregration(
    spatRaster = Data_rast,
    tResolution = "day",
    tStep = 2,
    fun = mean
)
}
\author{
Erik Kusch
}
