% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal_Decumulation.r
\name{Temporal_Decumulation}
\alias{Temporal_Decumulation}
\title{Carry out decumulation of temporally cumulatively stored data}
\usage{
Temporal_Decumulation(
  spatRaster,
  interval,
  cumulMode,
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{spatRaster}{A SpatRaster within which coverage should be identified}

\item{interval}{Number of layers in the spatRaster belonging to the same interval of cumulation.}

\item{cumulMode}{Character. cumulMode of cumulative storage of values. Currently supported: CDS (first layer per day is the cumulative sum of the previous day).}

\item{cores}{Integer. Number of cores for parallelisation if desired.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal cumulation and reverses the cumulative measurements into indvidual ones. Currently conceptualised to decumulate hourly CDS records.
}
\examples{
## Loading some data
Data_rast <- terra::rast(system.file("extdata", "KiN_Rast.nc", package = "ClimHub"))
## cropping to a smaller area for the example
Data_rast <- terra::crop(Data_rast, c(0, 7e4, 6.7e6, 6.77e6))
## creating a decumulation example dataset
Decumul_rast <- c(Data_rast, Data_rast, Data_rast, Data_rast,Data_rast[[1:8]])
## decumulation examples
# single-core
SingleCore <- Temporal_Decumulation(
    spatRaster = Decumul_rast,
    interval = 24,
    cumulMode = "CDS"
)
# multi-core
MultiCore <- Temporal_Decumulation(
    spatRaster = Decumul_rast,
    interval = 24,
    cumulMode = "CDS",
    cores = 4
)
## both are the same
all.equal(SingleCore, MultiCore)
}
\author{
Erik Kusch
}
