% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viz_SpatRast.r
\name{Viz_SpatRast}
\alias{Viz_SpatRast}
\title{Visualise spatraster data and overlay sf polygons/points if desired.}
\usage{
Viz_SpatRast(
  spatRaster,
  colour = viridis::inferno(100),
  legendTitle,
  panelLabels,
  panelColumns = 1,
  sfObj,
  sfSize = 1,
  sfShape = 1,
  sfColour = "black",
  sfFill = "NA"
)
}
\arguments{
\item{spatRaster}{A spatRaster object to visualise.}

\item{colour}{Colour palette for spatRaster visualisation.}

\item{legendTitle}{Optional. Character vector of title for colour legend. By default, the content of the terra::longnames() and terra::units() fields of the supplied spatRaster object.}

\item{panelLabels}{Optional. Character vector of labels to apply to each layer of the SpatRast. By default, the content of the terra::time() field of the supplied spatRaster object.}

\item{panelColumns}{Number of columns for panel arrangement of plots (each layer of the supplied spatRaster is visualised as an individual panel).}

\item{sfObj}{Optional. sf object which to overlay.}

\item{sfSize}{Optional. Size of sf overlay.}

\item{sfShape}{Optional. Shape of sf overlay if points.}

\item{sfColour}{Optional. Colour of sf overlay.}

\item{sfFill}{Optional. Fill colour of sf overlay.}
}
\value{
A ggplot2 object visualising a raster.
}
\description{
Uses the ggplot2 engine to easily create visualisations of spatraster data and overlay sf polygon or point data if desired.
}
\examples{
## base plotting
spatRaster <- terra::rast(system.file("extdata", "Sognefjord_TN.nc", package = "ClimHub"))[[1:2]]
Viz_SpatRast(spatRaster = spatRaster)

## changing colour and layout
p <- Viz_SpatRast(spatRaster = spatRaster, colour = viridis::viridis(1e3), panelColumns = 2)
p + ggplot2::theme(legend.position = "bottom")

## assigning custom panel and legend labels
p <- Viz_SpatRast(spatRaster = spatRaster, legendTitle = "Air Temperature [K]", panelLabels = c("First Panel", "Second Panel"), panelColumns = 2)
p + ggplot2::theme(legend.position = "bottom")

## adding an sf overlay
data("Jotunheimen_sf")
sfObj <- Jotunheimen_sf
Viz_SpatRast(spatRaster = spatRaster, sfObj = sfObj)

## larger area with more customisation
spatRaster <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]] / 10 # needs to be divided by 10 to get actual Kelvin data
Viz_SpatRast(spatRaster = spatRaster, legendTitle = "Mean Air Temperature [K]", sfObj = Jotunheimen_sf, sfColour = "white", sfSize = 0.5) +
    ggplot2::theme_dark() +
    ggplot2::theme(
        legend.position = "top",
        legend.key.width = unit(1.5, "cm"),
        legend.key.height = unit(1, "cm")
    ) +
    ggplot2::guides(
        fill =
            ggplot2::guide_colorbar(
                title.position = "top",
                title.hjust = 0.5
            )
    )
}
\author{
Erik Kusch
}
