% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRead.FileCheck.r
\name{WriteRead.FileCheck}
\alias{WriteRead.FileCheck}
\title{Checking if a file already exists}
\usage{
WriteRead.FileCheck(FName, Dir = getwd(), loadFun, load = TRUE, verbose = TRUE)
}
\arguments{
\item{FName}{File name}

\item{Dir}{Directory where to look for file}

\item{loadFun}{function with which to load filetype of FName}

\item{load}{Logical. Whether to load the data or not}

\item{verbose}{Logical. Whether to print/message function progress in console or not.}
}
\value{
Either a data object or NULL
}
\description{
If a file already exists in a given place, load that file.
}
