% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteRead.NC.r
\name{WriteRead.NC}
\alias{WriteRead.NC}
\title{Read or write NetCDF files and their metadata}
\usage{
WriteRead.NC(
  NC,
  FName,
  Variable,
  LongVar,
  Unit = "NA",
  Attrs = NULL,
  Write = FALSE,
  Compression = 1
)
}
\arguments{
\item{NC}{SpatRaster}

\item{FName}{Filename including directory}

\item{Variable}{Character. Needed for writing. Variable name to be saved as varname in NC output.}

\item{LongVar}{Character. Needed for writing. Long variable name.}

\item{Unit}{Character. Needed for writing. Unit in which Variable is recorded.}

\item{Attrs}{Named vector of metadata attributes to be written. Needed for writing.  When reading, all attributes in the .nc file are read.}

\item{Write}{Logical. Whether to write metadata}

\item{Compression}{Integer between 1 to 9. Needed for writing. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}
}
\value{
A SpatRaster with metadata
}
\description{
Read or write netcdf files and their metadata attributes.
}
