% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Access_KlimaiNorge2100.r
\name{Access_KlimaiNorge2100}
\alias{Access_KlimaiNorge2100}
\title{Download Klima i Norge 2100 data from the Norwegian Meteorological Institute}
\usage{
Access_KlimaiNorge2100(
  variable,
  dateStart,
  dateStop,
  model,
  scenario = c("rcp45", "rcp85"),
  cores = 1,
  fileName,
  compression = NA,
  removeTemporary = TRUE,
  writeFile = TRUE
)
}
\arguments{
\item{variable}{Character. An overview of Klima i Norge variables can be obtained with \code{Discovery_Variables(dataSet = "KlimaiNorge2100")}.}

\item{dateStart}{Character. "YYYY-MM-DD" date at which to start time series of downloaded data. Data is available daily at hourly intervals.}

\item{dateStop}{Character. "YYYY-MM-DD" date at which to stop time series of downloaded data. Data is available daily at hourly intervals.}

\item{model}{Character. An overview of climate models from which data can be obtained can be obtained with \code{Discovery_QuickFacts("KlimaiNorge2100")$models}.}

\item{scenario}{Character. An overview of climate models from which data can be obtained can be obtained with \code{Discovery_QuickFacts("KlimaiNorge2100")$scenarios}. Note that this choice only affects data post-dating 2005-12-31.}

\item{cores}{Optional, Integer. Number of cores to use for parallel downloads. Default NULL defines no parallelisation.}

\item{fileName}{Character. A file name for the produced file, including path.}

\item{compression}{Optional, Integer. Compression level between 1 to 9 applied to final .nc file. Same as compression argument in terra::writeCDF(). Defaults to NA.}

\item{removeTemporary}{Optional, Logical. Whether to delete temporary files after completion. Defaults to TRUE.}

\item{writeFile}{Optional, Logical. Whether to write final SpatRaster to disk as an .nc or to return information from memory. Setting to FALSE will prohibit removal of temporary files. Defaults to TRUE.}
}
\value{
SpatRaster. Contains the downloaded data and metadata attributes that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string for in-line citation of the data product
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call
}
}
\description{
Downloads and processes data from the \href{https://www.miljodirektoratet.no/publikasjoner/2015/september-2015/klima-i-norge-2100/}{Klima_i_Norge_2100} data product hosted through \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/catalog.html}{thredds.met.no}.
Specifically, this function provides access to the following datasets:
\enumerate{
\item \href{https://publikasjoner.nve.no/rapport/2016/rapport2016_59.pdf}{Gridded 1 x 1 km climate and hydrological projections for Norway} data contained within \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/utgave2015/catalog.html}{utgave 2015}.
}
}
\examples{
\dontrun{
Access_KlimaiNorge2100(
    variable = "Maximum Air Temperature",
    dateStart = "2004-08-01",
    dateStop = "2006-09-17",
    model = "CNRM_CCLM",
    scenario = "rcp85",
    cores = 1,
    fileName = "KlimaiNorge2100.nc",
    compression = 9,
    removeTemporary = TRUE
)
}
}
\author{
Erik Kusch
}
