% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Access_NORA3.r
\name{Access_NORA3}
\alias{Access_NORA3}
\title{Access NORA3 3km Norwegian Reanalysis Data}
\usage{
Access_NORA3(
  variable,
  dateStart,
  dateStop,
  leadTimeHour,
  fileName,
  compression = NA,
  downloadData = TRUE,
  writeFile = TRUE
)
}
\arguments{
\item{variable}{Character. An overview of NORA3 variables can be obtained with \code{Discovery_Variables(dataSet = "NORA3")}.}

\item{dateStart}{Character. "YYYY-MM-DD HH" date at which to start time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{dateStop}{Character. "YYYY-MM-DD HH" date at which to stop time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{leadTimeHour}{Integer. Lead time of reanalysis. NORA3 leadtimes can be obtained with \code{Discovery_QuickFacts("NORA3")$leadtime}.}

\item{fileName}{Character. A file name for the produced file, including path.}

\item{compression}{Optional, Integer. Compression level between 1 to 9 applied to final .nc file. Same as compression argument in terra::writeCDF(). Defaults to NA.}

\item{writeFile}{Optional, Logical. Whether to write final SpatRaster to disk as an .nc or to return information from memory. Setting to FALSE will prohibit removal of temporary files. Defaults to TRUE.}

\item{removeTemporary}{Optional, Logical. Whether to delete temporary files after completion. Defaults to TRUE.}
}
\value{
SpatRaster. Contains the downloaded data and metadata attributes that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string for in-line citation of the data product
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call
}
}
\description{
Downloads and processes data from the NORA3 data product hosted through \href{https://thredds.met.no/thredds/projects/nora3.html}{thredds.met.no}.
Specifically, this function provides access to the NORA3 files contained within \href{https://thredds.met.no/thredds/catalog/nora3/catalog.html}{nora3}.
}
\examples{
\dontrun{
NORA3 <- Access_NORA3(
    variable = "TS (Surface temperature)", # which variable
    dateStart = "1961-08-01 00", dateStop = "1961-08-02 18", # time-window
    leadTimeHour = 3
    fileName = "NORA3.nc", compression = 9 # file storing
)
unlink("NORA3.nc")
}
}
\author{
Erik Kusch
}
