% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Access_NORA3.r
\name{Access_NORA3}
\alias{Access_NORA3}
\title{Access NORA3 3km Norwegian Reanalysis Data}
\usage{
Access_NORA3(
  variable,
  dateStart,
  dateStop,
  extent = NULL,
  leadTimeHour,
  fileName,
  compression = NA
)
}
\arguments{
\item{variable}{Character. An overview of NORA3 variables can be obtained with \code{Discovery_Variables(dataSet = "NORA3")}.}

\item{dateStart}{Character. "YYYY-MM-DD HH" date at which to start time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{dateStop}{Character. "YYYY-MM-DD HH" date at which to stop time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{extent}{Optional. The extent to subset the data to, in coordinates of the projection or decimal degrees of longitude and latitude. A numeric vector of length 4 with values minimum and maximum X/longitude and minimum and maximum Y/latitude, in that order. Valid ranges for projected X and Y coordinates are \verb{(778360.875, 3442360.75, -1270477, 3193523)}, for longitude/latitude values \verb{(-30.168  85.793  44.025  84.057)}. Defaults to \code{NULL}, returning full spatial range of data.}

\item{leadTimeHour}{Integer. Lead time of reanalysis. NORA3 lead times can be obtained with \code{Discovery_QuickFacts("NORA3")$leadtime}.}

\item{fileName}{Character. A file name for the produced file, including path.}

\item{compression}{Optional, integer. Compression level between 1 to 9 applied to final netCDF file. Defaults to NA (no compression applied). Currently not used due to ncdfCF saving scheme.}
}
\value{
A \code{CFDataset} object which contains the downloaded data and relevant metadata attributes. If specified, also writes a netCDF file to disk.
}
\description{
Downloads and processes data from the NORA3 data product hosted through \href{https://thredds.met.no/thredds/projects/nora3.html}{thredds.met.no}.
Specifically, this function provides access to the NORA3 files contained within \href{https://thredds.met.no/thredds/catalog/nora3/catalog.html}{nora3}.
}
\examples{
\dontrun{
NORA3 <- Access_NORA3(
  variable = "TS", # which variable
  dateStart = "1961-08-01 00", dateStop = "1961-08-02 18", # time-window
  extent = c(0, 40, 50, 60),
  leadTimeHour = 3,
  fileName = "NORA3.nc", compression = 9 # file storing
)
}
}
\author{
Erik Kusch, Patrick Van Laake
}
