% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.KlimaINorge2100.r
\name{Download.KlimaINorge2100}
\alias{Download.KlimaINorge2100}
\title{Downloading Klima i Norge 2100 Data from the Norwegian Meteorological Instite}
\usage{
Download.KlimaINorge2100(
  Variable = "TM",
  DataSet = "utgave2015",
  Model = "MPI_RCA",
  Scenario = "hist",
  File = NULL,
  DateStart = 1971,
  DateStop = 2005,
  Dir = getwd(),
  FileName,
  Compression = 9
)
}
\arguments{
\item{Variable}{Character. Ignored when Dataset != "utgave2015". Klima_i_Norge_2100 contains the following:
\itemize{
\item "TX" ... Daily_maximum_air_temperature
\item "TN" ... Daily_minimum_air_temperature
\item "TM" ... Daily_mean_air_temperature
\item "SWE" ... Snow_water_equivalent_amount
\item "SMD" ... Soil_moisture_deficit_amount_per_day
\item "RUN" ... runoff_amount_per_day
\item "RR" ... precipitation_amount_per_day
\item "GRW" ... groundwater_amount
\item "ORO" ... orography
}}

\item{DataSet}{Character. Which dataset to query data from. Supported strings:
\itemize{
\item "utgave2015" (for annual 1x1km historic and projection data)
\item "nyindekser2015" (for 1x1km climate indices)
\item "masks" (for grids of regions and counties across Norway)
}}

\item{Model}{Character. Ignored when Dataset != "utgave2015". Klima_i_Norge_2100 contains the following:
\itemize{
\item CNRM_CCLM
\item CNRM_RCA
\item EC-EARTH_CCLM
\item EC-EARTH_HIRHAM
\item EC-EARTH_RACMO
\item EC-EARTH_RCA
\item HADGEM_RCA
\item IPSL_RCA
\item MPI_CCLM
\item MPI_RCA.
}}

\item{Scenario}{Character. Ignored when Dataset != "utgave2015". Klima_i_Norge_2100 contains the following:
\itemize{
\item rcp85 ... RCP85 (2005 - 2100)
\item "rcp45" ... RCP45 (2005 - 2100)
\item "hist" ... Historical conditions (1971 - 2005)
}}

\item{File}{Character. Required when Dataset != "utgave2015". Name of .nc file hosted on thredds.}

\item{DateStart}{Numeric. Ignored when Dataset != "utgave2015". Year ('YYYY') at which to start time series of downloaded data.}

\item{DateStop}{Numeric. Ignored when Dataset != "utgave2015". Year ('YYYY') at which to stop time series of downloaded data.}

\item{Dir}{Character/Directory Pointer. Directory specifying where to download data to.}

\item{FileName}{Character. A file name for the produced file.}

\item{Compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}
}
\value{
A SpatRaster object containing the downloaded data, and a file in the specified directory. The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string which to use for in-line citation of the data product obtained}.
}
}
\description{
This function is used to obtain the \href{https://www.miljodirektoratet.no/publikasjoner/2015/september-2015/klima-i-norge-2100/}{Klima_i_Norge_2100} data product hosted through \href{https://thredds.met.no/thredds/catalog/KSS/catalog.html}{thredds.met.no}. Specifically, this function makes available the following datasets:
\enumerate{
\item \href{https://publikasjoner.nve.no/rapport/2016/rapport2016_59.pdf}{Gridded 1 x 1 km climate and hydrological projections for Norway} data contained within \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/utgave2015/catalog.html}{utgave 2015}.
\item \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/nyeindekser2015/catalog.html}{1x1km Climate Indices}
\item \href{https://thredds.met.no/thredds/catalog/KSS/Klima_i_Norge_2100/masks/catalog.html}{Grids of Regions and Counties across Norway}
}
}
\examples{
\dontrun{
UtgaveTest <- Download.KlimaINorge2100(
  Variable = "TM", # which variable
  DataSet = "utgave2015", # data set
  Model = "MPI_RCA", # climate model
  Scenario = "hist", # scenario
  File = NULL, # single file specification
  DateStart = 1971, DateStop = 2005, # time-window
  Dir = getwd(), FileName = "TM-HIST_MPI_RCA" # file storing
)
terra::plot(UtgaveTest[[1:2]])

UtgaveORO <- Download.KlimaINorge2100(
  Variable = "ORO",
  DataSet = "utgave2015",
  FileName = "ORO"
)

SingleFile <- Download.KlimaINorge2100(
  DataSet = "nyeindekser2015",
  File = "snowMap1x1.nc",
  FileName = "SnowIndex"
)
}
}
