% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Download.NORA3.r
\name{Download.NORA3}
\alias{Download.NORA3}
\title{Downloading NORA3 3km NOrwegian Reanalysis Data from the Norwegian Meteorological Instite}
\usage{
Download.NORA3(
  Variable = "TS (Surface temperature)",
  DateStart = "1961-08-01 00",
  DateStop = "2022-12-31 18",
  Leadtime = 3,
  Cores = 1,
  Dir = getwd(),
  FileName,
  Compression = 9,
  RemoveTemporary = TRUE
)
}
\arguments{
\item{Variable}{Character. An overview of NORA3 variables can be obtained with Meta.Variables(dataset = "NORA3").}

\item{DateStart}{Character. "YYYY-MM-DD HH" date at which to start time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{DateStop}{Character. "YYYY-MM-DD HH" date at which to stop time series of downloaded data. Data is available daily at hours 00, 06, 12, and 18.}

\item{Leadtime}{Integer. Lead time of reanalysis, either 3, 4, 5, 6, 7, 8 or 9.}

\item{Cores}{Integer. How many cores to use for parallel downloads. Default NULL defines no parallelisation.}

\item{Dir}{Character/Directory Pointer. Directory specifying where to download data to.}

\item{FileName}{Character. A file name for the produced file.}

\item{Compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}

\item{RemoveTemporary}{Logical. Whether to delete temporary files after completion.}
}
\value{
A SpatRaster object containing the downloaded data, and a file in the specified directory. The SpatRaster contains metadata/attributes as a named vector that can be retrieved with terra::metags(...):
\itemize{
\item \emph{Citation} - A string which to use for in-line citation of the data product obtained.
\item \emph{Call_} - A set of strings matching arguments supplied to the download function call.
}
}
\description{
This function is used to obtain the NORA3 data product hosted through \href{https://thredds.met.no/thredds/projects/nora3.html}{thredds.met.no}. Specifically, this function makes available the following datasets:
\enumerate{
\item NORA3 files contained within \href{https://thredds.met.no/thredds/catalog/nora3/catalog.html}{nora3}.
}
}
\examples{
\dontrun{
NORA3 <- Download.NORA3(
    Variable = "TS (Surface temperature)", # which variable
    DateStart = "1961-08-01 00", DateStop = "1961-08-02 18", # time-window
    Leadtime = 3, Cores = 1,
    Dir = getwd(), FileName = "NORA3", Compression = 9, # file storing
    RemoveTemporary = TRUE
)
}
}
