% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Handle.ListLayers.r
\name{Handle.ListLayers}
\alias{Handle.ListLayers}
\title{Layer-wise application on SpatRasters in a list}
\usage{
Handle.ListLayers(Rast_ls, FUN = mean, verbose = TRUE)
}
\arguments{
\item{Rast_ls}{List. List of SpatRasters whose layers to compute with each other. All SpatRasters need to have the same layer count.}

\item{FUN}{Function. Function by which to summarise/compute SpatRaster layers with each other.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Loops over layers of SpatRasters within supplied list object and calculates one resulting layer per layer-group across SpatRasters in accordance with supplied FUN argument.
}
\examples{
## load the first 2 days of the KiN temperature data provided with ClimHub to two separate list slots and articifially increase values by 1 for the second list item
Rast_ls <- list(
    Base = terra::rast(system.file("extdata", "KiN_AT.nc", package = "ClimHub"))[[1:2]],
    Base1 = terra::rast(system.file("extdata", "KiN_AT.nc", package = "ClimHub"))[[1:2]]+1
    )
## calculate difference between each layer in the two list slots
Diff_rast <- Handle.ListLayers(Rast_ls = Rast_ls, FUN = diff)
Diff_rast # all values in this 2-layer raster, ought to be 1 or NA

}
