% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.LoadFiles.r
\name{Helper.LoadFiles}
\alias{Helper.LoadFiles}
\title{Load list of NetCDF files and stack them into a SpatRaster}
\usage{
Helper.LoadFiles(FNames)
}
\arguments{
\item{FNames}{Character. Vector of absolute filenames to be loaded.}
}
\value{
A spatraster object.
}
\description{
Iterates of a number of filenames, pointing to NetCDFs, and loads them into one SpatRaster.
}
