% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_InputChecker.r
\name{Helper_InputChecker}
\alias{Helper_InputChecker}
\title{Check Validity of User Input}
\usage{
Helper_InputChecker(inputCheck)
}
\arguments{
\item{inputCheck}{Optional, list. A named list of lists where each sublist contains three elements: Input (the value to check),
Allowed (the allowed values or range), and Operator (either "in" for membership or "exceeds" for range checks).}
}
\value{
Nothing. Only generates an error if a check did not succeed.
}
\description{
Loops over contents of a named list of lists where each sublist contains the pointers Input, Allowed, and Operator.
}
\examples{
Variable <- "Maximum Air Temperature"
DateStart <- "1971-01-01"
DateStop <- "1971-08-31"
Model <- "CNRM_CCLM"
Scenario <- "rcp85"

Start <- as.POSIXct(paste0(DateStart, "00:00:00"), tz = "UTC")
Stop <- as.POSIXct(paste0(DateStop, "00:00:00"), tz = "UTC")

inputCheck <- list(
    Variable = list(
        Input = Variable,
        Allowed = Discovery_Variables("KlimaiNorge2100")$name,
        Operator = "in"
    ),
    Time = list(
        Input = c(Start, Stop),
        Allowed = Discovery_QuickFacts("KlimaiNorge2100")$time$extent,
        Operator = "exceeds"
    ),
    Models = list(
        Input = Model,
        Allowed = Discovery_QuickFacts("KlimaiNorge2100")$models,
        Operator = "in"
    ),
    Scenarios = list(
        Input = Scenario,
        Allowed = Discovery_QuickFacts("KlimaiNorge2100")$scenarios,
        Operator = "in"
    )
)

Helper_InputChecker(inputCheck)
}
\author{
Erik Kusch
}
