% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Threshold.r
\name{Helper_Threshold}
\alias{Helper_Threshold}
\title{Apply tresholding to raster layers}
\usage{
Helper_Threshold(
  spatRaster,
  operator,
  threshold,
  returnValues = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spatRaster}{A SpatRaster within which thresholding should be applied}

\item{operator}{Character. One of ">", "<", ">=", "<=", "==", "!=" to evaluate spatRaster cells against threshold value.}

\item{threshold}{Numeric or Character. threshold to evaluate spatRaster cells against using operator value.}

\item{returnValues}{Logical. Whether to return spatRaster of values matching thresholding (TRUE) or only logical raster indicating where thresholding is matched (FALSE).}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Loops over layers in a SpatRaster object and evaluates threshold criteria defined by user.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1:31]]
Helper_Threshold(
    spatRaster = Data_rast,
    operator = "<",
    threshold = 2650
)
Helper_Threshold(
    spatRaster = Data_rast,
    operator = "<",
    threshold = 2650,
    returnValues = TRUE
)
}
\author{
Erik Kusch
}
