% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics_ETCCDI.r
\name{Metrics_ETCCDI}
\alias{Metrics_ETCCDI}
\title{Calculate ETCCDI Indices from List Input}
\usage{
Metrics_ETCCDI(
  projectionList,
  baseLineList,
  TResolution = "year",
  RRThreshold = 42,
  indices,
  fileName
)
}
\arguments{
\item{projectionList}{List. List of \code{CFVariable} objects. Names of elements must be "TX", "TN", and "RR", holding maximum and minimum daily air temperature and daily total precipitation, respectively. Note that these data must be recorded in "K", "K", and "mm", respectively.}

\item{baseLineList}{List. List of \code{CFDataset} objects. Names of elements must be "TX_Base", "TN_Base", and "RR_Base", holding lower and upper quantiles of maximum and minimum daily air temperature and daily total precipitation, respectively. Note that these data must be recorded in "K", "K", and "mm", respectively.}

\item{TResolution}{Character. Temporal resolution for calculation of ETCCDI. Supports "year" (default), "month" and "season".}

\item{RRThreshold}{Numeric. Custom threshold for daily precipiation in mm for calculation of Rnnmm. Defaults to 42.}

\item{indices}{Optional, character. Character vector of ETCCDI abbreviations to calculate. If missing, all supported indices in this function are calculated.}

\item{fileName}{Character, optional. Character. A file name for the produced file, including path and ".nc" file ending. If no value is supplied, the dataset is not written to disk but returned as a \code{CFDataset} object in memory. If a file name is supplied and a file with that name already exists, the function will attempt to load and return that file instead of recalculating the indices.}
}
\value{
A \code{CFDataset} containing a CFVariable for each ETCCDI. Each variable is named by its ETCCDI acronym and has a \code{long_name} attribute describing the index.
}
\description{
This function calculates \href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/jgrd.50203}{ETCCDIs} from a named list of CFVariable objects. Currently, only some ETCCDI are supported.
}
\details{
Describe ETCCDI here including their abbreviations, titles, verbatim descriptions, units, and default temporal resolution. Also describe which data inputs are required for which index.
}
\examples{
\dontrun{

}
}
\author{
Erik Kusch
}
