% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics_ETCCDI.r
\name{Metrics_ETCCDI}
\alias{Metrics_ETCCDI}
\title{Calculate ETCCDI Indices from List Input}
\usage{
Metrics_ETCCDI(ls)
}
\arguments{
\item{ls}{Optional, list. List of SpatRasters. Names of elements must be "TX" and "TN", holding maximum and minimum daily air temperature respectively.}
}
\value{
A named list of SpatRasters with each element corresponding to an ETCCDI.
}
\description{
This function calculates \href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/jgrd.50203}{ETCCDIs} from a named list of SpatRaster objects. Currently, only the first four ETCCDI are supported.
}
\examples{
\dontrun{
TX_rast <- terra::rast("inst/extdata/Jotunheimen_TX.nc") # terra::rast(system.file("extdata", "Jotunheimen_TX.nc", package = "ClimHub"))
TN_rast <- terra::rast("inst/extdata/Jotunheimen_TN.nc")
RR_rast <- terra::rast("inst/extdata/Jotunheimen_RR.nc")
terra::units(RR_rast) <- "mm"
# BASEPeriod_rast <- terra::rast("inst/extdata/Jotunheimen_BASEPeriod.nc")

Metrics_ETCCDI(ls = list(TX = TX_rast, TN = TN_rast, RR = RR_rast
# , BASEPeriod = BASEPeriod_rast)
)
}

}
\author{
Erik Kusch
}
