% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrics_ETCCDI.r
\name{Metrics_ETCCDI}
\alias{Metrics_ETCCDI}
\title{Calculate ETCCDI Indices from List Input}
\usage{
Metrics_ETCCDI(projectionList)
}
\arguments{
\item{projectionList}{List. List of \code{CFVariable} objects. Names of elements must be "TX", "TN", and "RR", holding maximum and minimum daily air temperature and daily total precipitation, respectively. Note that these data must be recorded in "K", "K", and "mm", respectively.}
}
\value{
A named list of \code{CFVariable} objects with each element corresponding to an ETCCDI.
}
\description{
This function calculates \href{https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/jgrd.50203}{ETCCDIs} from a named list of CFVariable objects. Currently, only some ETCCDI are supported.
}
\examples{
\dontrun{
TX_CF <- NC_Read("inst/extdata/KiN_tx_2050.nc")[["tasmax"]]
TN_CF <- NC_Read("inst/extdata/KiN_tn_2050.nc")[["tasmin"]]
RR_CF <- NC_Read("inst/extdata/KiN_prc_2050.nc")[["pr"]] * 86400 # to get from mm/day to kg m-2 s-1
RR_CF$set_attribute("units", "NC_CHAR", "mm")

Metrics_ETCCDI(projectionList = list(TX = TX_CF, TN = TN_CF, RR = RR_CF))
}

}
\author{
Erik Kusch
}
