% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Read.r
\name{NC_Read}
\alias{NC_Read}
\title{Read a netCDF file}
\usage{
NC_Read(fileName, asTerra = TRUE)
}
\arguments{
\item{fileName}{Character. Fully qualified file name for a netCDF file on a local file system or the URL to a netCDF resource on a THREDDS server.}

\item{asTerra}{Logical. If \code{TRUE} (default), the contents of the netCDF file is returned as a \code{SpatRaster} object from the \code{terra} package.}
}
\value{
Either a \code{SpatRaster} object or an \code{ncdfCF} data set instance.
}
\description{
Open and read a netCDF file including its metadata.
}
\examples{
## loading as ncdfCF object
NC_Read(fileName = system.file("extdata", "KiN_rast.nc", package = "ClimHub"), asTerra = FALSE)
## loading as terra SpatRaster object
NC_Read(fileName = system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
}
\author{
Patrick Van Laake, Erik Kusch
}
