% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_Read.r
\name{NC_Read}
\alias{NC_Read}
\title{Read NetCDF files and their metadata}
\usage{
NC_Read(fileName)
}
\arguments{
\item{fileName}{Character. Filename including directory for reading/writing.}
}
\value{
A SpatRaster with metadata
}
\description{
Read NetCDF files and their metadata attributes. Metadata from the NetCDF file is read and attached to the returned SpatRaster via \code{terra::metags()}.
}
\examples{
Read_ras <- NC_Read(fileName = system.file("extdata", "KiN_rast.nc", package = "ClimHub"), meta = c("Citation"))
terra::metags(Read_ras)[terra::metags(Read_ras)$name == "Citation", ]
}
\author{
Erik Kusch
}
