% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Extract.r
\name{Spatial_Extract}
\alias{Spatial_Extract}
\title{Extracting data from SpatRaster with a SpatialFeaturesObject}
\usage{
Spatial_Extract(spatRaster, sf, fun = mean)
}
\arguments{
\item{spatRaster}{A SpatRaster within which coverage should be identified}

\item{sf}{Either an sf polygon(-collection) or an sf point(-collection).}

\item{fun}{User-defined function by which to aggregate values of cells within a polygon. Supported functions are mean, sum, min, max and table.}
}
\value{
A data.frame.
}
\description{
Extract data from SpatRaster corresponding to spatial features in sf object.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
data(Jotunheimen_sf)
Spatial_Extract(
    spatRaster = Data_rast,
    sf = Jotunheimen_sf,
    fun = mean
)
data(Nor2K_sf)
Spatial_Extract(
    spatRaster = Data_rast,
    sf = Nor2K_sf
)
}
\author{
Erik Kusch
}
