% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Limit.r
\name{Spatial_Limit}
\alias{Spatial_Limit}
\title{Cropping & Range Masking with Edge Support}
\usage{
Spatial_Limit(spatRaster, shape, verbose = TRUE)
}
\arguments{
\item{spatRaster}{A SpatRaster within which coverage should be identified}

\item{shape}{Either a SpatExtent or an sf polygon(-collection) whose coverage of the raster object is to be found.}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster.
}
\description{
Cropped and masking the original SpatRaster (\code{spatRaster}) using supplied SpatExtent or shapefile (\code{shape}) and retaining all pixels which are even just partially covered.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1:31]]
data(Jotunheimen_sf)
Data_rast <- Spatial_Reproject(Data_rast, Jotunheimen_sf)
Spatial_Limit(Data_rast, Jotunheimen_sf)
}
\author{
Erik Kusch
}
