% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_Reproject.r
\name{Spatial_Reproject}
\alias{Spatial_Reproject}
\title{Reprojecting spatial data.}
\usage{
Spatial_Reproject(projFrom, projTo, rasterResample = TRUE)
}
\arguments{
\item{projFrom}{Either a SpatRaster or SF object which is to be reprojected.}

\item{projTo}{Prefably, a character string identifying desired EPSG ID. Can also be a SpatRaster or SF object to whose projection the other opbject is to be reprojected. The latter specifications are less stable.}

\item{rasterResample}{Logical. If TRUE and both objects are SpatRasters, projFrom is also resampled to match the resolution and origin of ProjTo.}
}
\value{
A SpatRaster or SF object reprojected to the same CRS as in ProjTo.
}
\description{
Reprojecting one spatial data product to match another.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))[[1]]
## reprojecting raster
data(Jotunheimen_sf)
Spatial_Reproject(Data_rast, Jotunheimen_sf)
## reprojecting sf
data(Nor2K_sf)
Spatial_Reproject(Nor2K_sf, Data_rast)
}
\author{
Erik Kusch
}
