% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temporal.Aggregration.r
\name{Temporal.Aggregration}
\alias{Temporal.Aggregration}
\title{Carry out temporal aggregation}
\usage{
Temporal.Aggregration(Raster, TResolution, TStep, FUN, verbose = TRUE)
}
\arguments{
\item{Raster}{A SpatRaster within which coverage should be identified}

\item{TResolution}{Character. User-specified temporal resolution}

\item{TStep}{Numeric. User-specified time step}

\item{FUN}{User-defined aggregation function}

\item{verbose}{Logical. If progress should be displayed in the console.}
}
\value{
A SpatRaster
}
\description{
Takes a SpatRaster and user-specifications of temporal aggregation and carries it out.
}
\examples{
Data_rast <- terra::rast(system.file("extdata", "KiN_rast.nc", package = "ClimHub"))
Temporal.Aggregration(
    Raster = Data_rast,
    TResolution = "day",
    TStep = 2,
    FUN = mean
)
}
