% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteNC.r
\name{WriteNC}
\alias{WriteNC}
\title{Write NetCDF files and their metadata with a progress bar for netcdf layers}
\usage{
WriteNC(
  spatraster,
  output_file,
  compression = NA,
  variable,
  longname,
  unit,
  attrs = NULL
)
}
\arguments{
\item{spatraster}{SpatRaster}

\item{output_file}{Filename including directory}

\item{compression}{Integer between 1 to 9. Applied to final .nc file that the function writes to hard drive. Same as compression argument in terra::writeCDF().}

\item{variable}{Character. Variable name to be saved as varname in NC output.}

\item{longname}{Character. Long variable name.}

\item{unit}{Character. Unit in which Variable is recorded.}

\item{attrs}{Named vector of metadata attributes}
}
\value{
A SpatRaster with metadata written to the disk
}
\description{
Write NetCDF files and their metadata with a progress bar for netcdf layers. Adapted from the .writeCDF method in the terra package: https://github.com/rspatial/terra/blob/master/R/ncdf.R
}
